% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_accessors.R
\docType{methods}
\name{origin}
\alias{origin}
\alias{origin,antsImage-method}
\alias{origin<-,antsImage-method}
\title{Extract Image Attributes \code{origin} from \code{antsImage} or \code{nifti}
objects}
\usage{
origin(object)

\S4method{origin}{antsImage}(object)

\S4method{origin}{antsImage}(object) <- value
}
\arguments{
\item{object}{is an object of class \code{antsImage} or \code{nifti}.}

\item{value}{is the value to assign to the slot.}
}
\description{
Methods that act on the \code{slots} of the
\code{antsImage} or \code{nifti} object
}
