% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ants.extract}
\alias{ants.extract}
\title{Extract values from antsImage from mask}
\usage{
ants.extract(img, mask)
}
\arguments{
\item{img}{image either in array or class \code{antsImage}}

\item{mask}{binary image either in array or class \code{antsImage}}
}
\value{
Numeric vector
}
\description{
Just helpful wrapper to convert to array and then take
values for which mask is 1
}

