% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ants.overlay.R
\name{ants.overlay}
\alias{ants.overlay}
\title{Plot Lightbox for antsImage}
\usage{
ants.overlay(img, oimg, slices = 53, col.x = gray(0:64/64),
  col.y = "#FF0000FF", ...)
}
\arguments{
\item{img}{Image to be plotted}

\item{oimg}{image to be overlaid}

\item{slices}{slice to be displayed}

\item{col.x}{grayscale coloring for img}

\item{col.y}{color for overlay}

\item{...}{arguments to be passed to \code{\link[graphics]{image} }}
}
\description{
Plots series of Images for slices of antsImage
}

