% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/oro2ants.R
\name{ants2oro}
\alias{ants2oro}
\title{Convert Between antsImage and oro.nifti}
\usage{
ants2oro(img, reorient = FALSE)
}
\arguments{
\item{img}{Object of class antsImage}

\item{reorient}{Reorientation passed to \code{\link{readNIfTI}}}
}
\value{
Object of class \code{nifti}
}
\description{
NIfTI data can be converted between \code{antsImage}
(from the ANTsR package) and nifti S4 objects.
}

