% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_apply_transforms.R
\name{ants_apply_transforms}
\alias{ants_apply_transforms}
\title{Apply ANTs transforms to nifti images}
\usage{
ants_apply_transforms(fixed, moving, ...)
}
\arguments{
\item{fixed}{image defining domain into which the moving image is transformed.}

\item{moving}{image to be mapped to fixed space.}

\item{...}{Arguments to be passed to \code{\link{antsApplyTransforms}}.}
}
\value{
Output \code{\link{nifti}} image
}
\description{
Wraps \code{\link{antsApplyTransforms}} but allows for 
\code{\link{nifti}} objects from the \code{oro.nifti} package
}

