% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_bwlabel.R
\name{ants_bwlabel}
\alias{ants_bwlabel}
\title{ANTs BWLabel Clusters of Certain Size}
\usage{
ants_bwlabel(img, k = 1, binary = TRUE)
}
\arguments{
\item{img}{input image}

\item{k}{Minimum cluster size needed}

\item{binary}{(logical) Should the result be binary or numbered with cluster.}
}
\value{
\code{nifti} object
}
\description{
Get Cluster of certain size from \code{\link{labelClusters}}
}

