% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/T1_Registration_function.R
\name{ants_regwrite}
\alias{ants_regwrite}
\title{Registration to a Template}
\usage{
ants_regwrite(filename, skull_strip = FALSE, n3correct = FALSE,
  correction = "N3", retimg = TRUE, outfile = NULL,
  template.file = file.path(fsldir(), "data", "standard",
  "MNI152_T1_1mm_brain.nii.gz"), interpolator = "Linear",
  other.files = NULL, other.outfiles = NULL, native.cereb = FALSE,
  native.fname = NULL, atlas.file = NULL, typeofTransform = "SyN",
  remove.warp = FALSE, outprefix = NULL, bet.opts = "-B -f 0.1 -v",
  betcmd = "bet", ...)
}
\arguments{
\item{filename}{filename of T1 image}

\item{skull_strip}{do skull stripping with FSL BET}

\item{n3correct}{do Bias correction}

\item{correction}{N3 or N4 correction, see \code{\link{bias_correct}}}

\item{retimg}{return a nifti object from function}

\item{outfile}{output filename should have .nii or .nii.gz
extension}

\item{template.file}{Filename of template to warp to}

\item{interpolator}{interpolation done for
\code{\link{antsApplyTransforms}}}

\item{other.files}{Filenames of other iamges to be
transformed with the T1}

\item{other.outfiles}{Output filenames of \code{other.files} to
be written}

\item{native.cereb}{Logical indicating if native cerebellum should be
created to \code{native.fname}}

\item{native.fname}{filename of native cerebellum file}

\item{atlas.file}{Filename of atlas used for warping labels}

\item{typeofTransform}{type of transformed used, passed to
\code{\link{antsRegistration}}}

\item{remove.warp}{(logical) Should warping images be deleted?}

\item{outprefix}{Character path of where the warp files should be stored.
Required if \code{remove.warp = FALSE}}

\item{bet.opts}{Options passed to \code{\link{fslbet}}}

\item{betcmd}{BET command used, passed to \code{\link{fslbet}}}

\item{...}{arguments to \code{\link{antsApplyTransforms}}}
}
\value{
NULL or object of class nifti for transformed T1 image
}
\description{
This function performs registration to a template
using ANTsR and SyN transformation, carry out the transformation
on other images, and can back-transform atlases
in template space to the native space of the iimage
}

