% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsCopy.R
\name{antsCopy}
\alias{antsCopy}
\alias{antsCopyDirection}
\alias{antsCopyOrigin}
\alias{antsCopySpacing}
\title{Copy Information of One Image to Another}
\usage{
antsCopyOrigin(reference, target)

antsCopyDirection(reference, target)

antsCopySpacing(reference, target)
}
\arguments{
\item{reference}{\code{antsImage} object to get origin values from}

\item{target}{\code{antsImage} object to copy origin values to}
}
\value{
\code{target} object of class \code{antsImage}
}
\description{
Wraps \code{antsSetOrigin/antsSetDirection/antsSetSpacing} 
and \code{antsGetOrigin/antsGetDirection/antsGetSpacing}
to copy the information over.  Finer control than 
\code{antsCopyImageInfo}
}
\examples{
library(ANTsR)
img <- makeImage(c(10,10),rnorm(100))
img2 <- makeImage(c(10,10), rnorm(100))
img2 <- antsCopyOrigin(img, img2)
}

