% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_multi_reg.R
\docType{methods}
\name{apply_multi_reg-methods}
\alias{apply_multi_reg-methods}
\alias{apply_multi_reg}
\alias{apply_multi_reg,list,list-method}
\alias{apply_multi_reg,character,list-method}
\alias{apply_multi_reg,antsImage,character-method}
\alias{apply_multi_reg,nifti,character-method}
\alias{apply_multi_reg,list,character-method}
\title{Check if antsImage or read in}
\usage{
apply_multi_reg(infiles, transformlist, fixed, ...)

\S4method{apply_multi_reg}{list,list}(infiles, transformlist, fixed, ...)

\S4method{apply_multi_reg}{character,list}(infiles, transformlist, fixed, ...)

\S4method{apply_multi_reg}{antsImage,character}(infiles, transformlist, fixed, ...)

\S4method{apply_multi_reg}{nifti,character}(infiles, transformlist, fixed, ...)
}
\arguments{
\item{infiles}{Input image files}

\item{transformlist}{list or character vector of transforms generated 
by antsRegistration where each transform is a filename.}

\item{fixed}{fixed image defining domain into which the infiles image 
is transformed}

\item{...}{Arguments passed to \code{\link{antsApplyTransforms}}}
}
\value{
Object of class \code{\link{nifti}}
}
\description{
Simple check to see if input is character, list, \code{nifti},
or class \code{antsImage}
}
\seealso{
\code{\link{antsApplyTransforms}}
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}
