% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bias_correct}
\alias{bias_correct}
\title{N3 or N4 Correct}
\usage{
bias_correct(file, correction = c("N3", "N4", "N4_Field"), outfile = NULL,
  retimg = FALSE, reorient = FALSE, shrinkfactor = "4", ...)
}
\arguments{
\item{file}{(character) image to be manipulated}

\item{correction}{(character) N3 or N4 correction?}

\item{outfile}{(character) resultant image name (optional)}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{shrinkfactor}{Shrink factor passed to
\code{\link{N3BiasFieldCorrection}}}

\item{...}{additional arguments passed to
\code{\link{N3BiasFieldCorrection}}}
}
\value{
If \code{retimg} then object of class nifti.  Otherwise,
Result from system command, depends if intern is TRUE or FALSE.
}
\description{
This function wraps ANTsR bias field corrections and returns
nifti objects
}

