% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_extract.R
\name{brain_extract}
\alias{brain_extract}
\title{Wrapper for ANTs brain extraction}
\usage{
brain_extract(
  img,
  template,
  template.mask,
  typeofTransform = c("SyNabp", "SyN"),
  correct = TRUE,
  reproducible = TRUE,
  seed = 1,
  ...
)
}
\arguments{
\item{img}{Image for brain extraction, nifti, character or antsImage}

\item{template}{Template image with mask, nifti, character or antsImage}

\item{template.mask}{Template brain mask}

\item{typeofTransform}{registration type: 'SyNabp' (better, slower), 
'SyN' (fast)}

\item{correct}{Should \code{\link{abpN4}} be performed first?}

\item{reproducible}{Sets the seed and 
\code{Sys.setenv(ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS = 1)}.
 See
\url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
for discussion.}

\item{seed}{will execute 
\code{Sys.setenv(ANTS_RANDOM_SEED = seed)} before
running to attempt a more reproducible result.   If \code{NULL}, will not set anything, 
but \code{reproducible} must be \code{FALSE}.}

\item{...}{additional arguments passed to \code{\link{abpBrainExtraction}}}
}
\value{
A list of the brain image, brain mask, k-means segmentation,
and transformations.
}
\description{
This function wraps \code{\link{abpN4}} and
\code{\link{abpBrainExtraction}} for \code{nifti} images with additional checks
}
