% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3d.R
\name{c3d}
\alias{c3d}
\title{Convert NeuroImage Types}
\usage{
c3d(infile, outfile, ...)
}
\arguments{
\item{infile}{input filename}

\item{outfile}{output filename with extension}

\item{...}{arguments to be passed to \code{\link{antsImageRead}}}
}
\value{
Output from \code{\link{antsImageWrite}}
}
\description{
Simple wrapper for \code{\link{antsImageRead}} and 
\code{\link{antsImageWrite}} for conversion of image file types. 
For example, NRRD files with nhdr/raw file formats can be converted to
\code{.nii.gz}.
}
