% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ants.R
\docType{methods}
\name{check_ants-methods}
\alias{check_ants}
\alias{check_ants,antsImage-method}
\alias{check_ants,character-method}
\alias{check_ants,list-method}
\alias{check_ants,nifti-method}
\alias{check_ants-methods}
\title{Check if antsImage or read in}
\usage{
check_ants(x, dimension = 3)

\S4method{check_ants}{antsImage}(x, dimension = 3)

\S4method{check_ants}{nifti}(x, dimension = 3)

\S4method{check_ants}{character}(x, dimension = 3)

\S4method{check_ants}{list}(x, dimension = 3)
}
\arguments{
\item{x}{character path of image or 
an object of class antsImage}

\item{dimension}{(numeric) passed to 
\code{\link{antsImageRead}} if the image
is read in}
}
\value{
antsImage object
}
\description{
Simple check to see if input is character, list, \code{nifti},
or class \code{antsImage}
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}
\seealso{
\code{\link{antsImageRead}}
}

