% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/check_ants.R
\name{check_ants}
\alias{check_ants}
\title{Check if antsImage or read in}
\usage{
check_ants(x, dimension = 3)
}
\arguments{
\item{x}{character path of image or
an object of class antsImage}

\item{dimension}{(numeric) passed to
\code{\link{antsImageRead}} if the image
is read in}
}
\value{
antsImage object
}
\description{
Simple check to see if input is character or of
class antsImage
}
\seealso{
\link{antsImageRead}
}

