% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nifti.R
\docType{methods}
\name{check_nifti,antsImage-method}
\alias{check_nifti,antsImage-method}
\title{Check if nifti image or read in a nifti image}
\usage{
\S4method{check_nifti}{antsImage}(x, reorient = FALSE,
  allow.array = FALSE, fast = FALSE, need_header = TRUE, ...)
}
\arguments{
\item{x}{character path of image or 
an object of class nifti, or array}

\item{reorient}{(logical) passed to \code{\link{readnii}} 
if the image
is to be re-oriented}

\item{allow.array}{(logical) Are array types allowed (TRUE) or
should there be an error if the object is not character or class
nifti.}

\item{fast}{not used for \code{antsImage} classes}

\item{need_header}{if \code{TRUE}, then an image type with header information
will be returned.  If not, then an array is fine.  Used really only in 
conjunction with \code{allow.array}}

\item{...}{additional options to pass to \code{\link{ants2oro}}}
}
\value{
nifti object or array if allow.array=TRUE and x is an array
}
\description{
Simple check to see if input is character or of 
class nifti
}
\seealso{
\link{readnii}
}
