% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkimg_antsImage.R
\docType{methods}
\name{checkimg-methods}
\alias{checkimg}
\alias{checkimg-methods}
\title{Force object to filename}
\usage{
\S4method{checkimg}{antsImage}(file, ...)
}
\description{
Ensures the output to be a character filename (or vector) from an input
image or \code{nifti}.
}

