% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_transform_to_file.R
\name{compose_transform_to_file}
\alias{compose_transform_to_file}
\title{Compose Transforms from ANTsR to a composite file}
\usage{
compose_transform_to_file(fixed, moving, transformlist = NULL,
  output_prefix = NULL)
}
\arguments{
\item{fixed}{fixed image defining domain into which the moving image is transformed.}

\item{moving}{moving image to be mapped to fixed space.}

\item{transformlist}{haracter vector of transforms generated by 
\code{\link{antsRegistration}} where each transform is a filename.}

\item{output_prefix}{Prefix of output filename to return.  The prefix will
add \code{comptx.nii.gz} to the end}
}
\value{
A character filename
}
\description{
This is a thin wrapper for \code{\link{antsApplyTransforms}} but 
is specifically for composing a transform and getting a file out.
}
\examples{
fixed <- antsImageRead( getANTsRData("r16") ,2)
moving <- antsImageRead( getANTsRData("r64") ,2)
fixed <- resampleImage(fixed,c(64,64),1,0)
moving <- resampleImage(moving,c(64,64),1,0)
mytx <- antsRegistration(fixed=fixed , moving=moving ,
                         typeofTransform = c("SyN") )
composed <- compose_transform_to_file( 
fixed=fixed, moving=moving, transformlist=mytx$fwdtransforms)
}
