% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_img.R
\name{corr_img}
\alias{corr_img}
\title{Neighborhood Correlation Image}
\usage{
corr_img(img1, img2, mask = NULL, radius = rep(1, 3))
}
\arguments{
\item{img1}{image of class \code{nifti}, \code{antsImage}, \code{character}}

\item{img2}{image of class \code{nifti}, \code{antsImage}, \code{character}}

\item{mask}{Binary image of class \code{nifti}, 
\code{antsImage}, \code{character}.  Only neighborhoods inside the mask 
will be taken}

\item{radius}{vector of length 3 for number of voxels to go in each direction.
Default is 27 neighbors (including voxel at center).  
Passed to \code{\link{getNeighborhoodInMask}}.}
}
\value{
Object of class \code{nifti}
}
\description{

}
\examples{
\dontrun{
corr_img("T1_Image.nii.gz", "FLAIR_Image.nii.gz", mask = "Mask.nii.gz")
}
}

