% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cort_thickness.R
\name{cort_thickness}
\alias{cort_thickness}
\alias{kk}
\title{Compute cortical thickness using Kelly Kapowski (ANTsR)}
\usage{
cort_thickness(seg, gray, white, ...)

kk(seg, gray, white, ...)
}
\arguments{
\item{seg}{segmentation image}

\item{gray}{gray matter probability image}

\item{white}{white matter probability image}

\item{...}{arguments passed to \code{\link{kellyKapowski}}}
}
\value{
Object of class \code{nifti}.
}
\description{
Wrapper for the \code{\link{kellyKapowski}} function
in ANTsR, that computes cortical thickness.
}

