% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_moment.R
\name{create_moment}
\alias{create_moment}
\title{Create Neighborhood Momemts of an Image}
\usage{
create_moment(img, mask = NULL, radius = rep(1, 3), retimg = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{img}{Object of class \code{nifti}, \code{character} or \code{antsImage}}

\item{mask}{Object of class \code{nifti}, \code{character} or \code{antsImage}.
Only neighborhoods inside the mask will be taken.  If \code{NULL}, all
neighbors of the image will be taken}

\item{radius}{array of values for neighborhood radius (in voxels)}

\item{retimg}{Should images (of class \code{nifti}) be returned or should
a matrix}

\item{verbose}{print diagnostic messages}

\item{...}{arguments to passed to \code{\link{neighborhood}}}
}
\value{
List of images (of class \code{nifti}) or
a matrix (V by 6), where V is the number of voxels in the mask
}
\description{
Creates first (mean) through fourth (kurtosis) central moments,
as well as gradients and mean/sd image.
}

