% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CT_Skull_Strip_robust.R
\name{CT_Skull_Strip_robust}
\alias{CT_Skull_Strip_robust}
\title{Robust CT Skull Stripping}
\usage{
CT_Skull_Strip_robust(img, outfile = NULL, keepmask = TRUE,
  maskfile = NULL, retimg = FALSE, reorient = FALSE, int = "0.01",
  lthresh = 0, uthresh = 100, nvoxels = 5, smooth.factor = 2,
  verbose = TRUE, opts = NULL, ...)
}
\arguments{
\item{img}{(character) File to be skull stripped or object of class
nifti}

\item{outfile}{(character) output filename}

\item{keepmask}{(logical) Should we keep the mask?}

\item{maskfile}{(character) Filename for mask (if \code{keepmask = TRUE}).
If \code{NULL}, then will do \code{paste0(outfile, "_Mask")}.}

\item{retimg}{(logical) return image of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{lthresh}{(default: 0) Lower value to threshold CT
\code{\link{fslthresh}}}

\item{uthresh}{(default: 100) Upper value to threshold CT
\code{\link{fslthresh}}}

\item{nvoxels}{Number of voxels to dilate/erode.  See \code{\link{dil_ero}}.}

\item{smooth.factor}{Smoothing factor for \code{\link{fslbet}}.  See \code{-w}
option in \code{fslbet.help()}.}

\item{verbose}{(logical) Should diagnostic output be printed?}

\item{opts}{Not used}

\item{...}{additional arguments passed to \code{\link{CT_Skull_Strip}}.}
}
\value{
Skull-stripped \code{nifti} object
}
\description{
Skull Stripping (using FSL's BET) a CT file using \code{fslr}
functions and robustified by registration
}

