% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_self.R
\name{diff_self}
\alias{diff_self}
\title{Flipped Difference Image}
\usage{
diff_self(img, swapdim = TRUE, verbose = TRUE, reproducible = TRUE,
  seed = 1)
}
\arguments{
\item{img}{Object of class \code{nifti}, \code{ANTsR}, or \code{character}}

\item{swapdim}{Should the image be reoriented to RPI then flipped, registered,
then changed back?}

\item{verbose}{Print diagnostic messages}

\item{reproducible}{Sets the seed and 
\code{Sys.setenv(ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS = 1)}.
 See
\url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
for discussion.}

\item{seed}{will execute 
\code{Sys.setenv(ANTS_RANDOM_SEED = seed)} before
running to attempt a more reproducible result.   If \code{NULL}, will not set anything, 
but \code{reproducible} must be \code{FALSE}.}
}
\value{
Object of class \code{nifti}
}
\description{
This function flips the image over the left/right axis, then registers 
the flipped image to the original image and then takes a difference
}
