% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_neck.R
\name{double_remove_neck}
\alias{double_remove_neck}
\title{Remove Neck Twice from a Scan}
\usage{
double_remove_neck(file, template.file, template.mask,
  typeofTransform = "Rigid", rep.value = 0, swapdim = TRUE,
  verbose = TRUE, ret_mask = FALSE)
}
\arguments{
\item{file}{File for neck removal - either filename or class nifti}

\item{template.file}{Template to warp to original image space}

\item{template.mask}{Mask of template to use as rough brain mask.  If
\code{template.file} is specified, but \code{template.mask} is not,
then \code{fslbin(file=template.file)} is performed.}

\item{typeofTransform}{Transformation for template to image, passed to
\code{\link{ants_regwrite}}.}

\item{rep.value}{Value to replace neck slices with}

\item{swapdim}{Should the dimensions be swapped before registration,
and then reset after}

\item{verbose}{Print out diagnostic messages}
}
\value{
Object of class nifti
}
\description{
Wrapper for running \code{remove_neck} twice, as some images
require this if large number of slices of the neck or upper shoulders
were scanned
}

