% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyImageDimensionsMask.R
\docType{methods}
\name{emptyImageDimensionsMask}
\alias{emptyImageDimensionsMask}
\alias{emptyImageDimensionsMask,nifti-method}
\title{Make Mask from Empty Image Dimensions}
\usage{
\S4method{emptyImageDimensionsMask}{antsImage}(img, ..., reorient = FALSE)
}
\arguments{
\item{img}{nifti object}

\item{...}{Arguments to be passed to \code{\link{getEmptyImageDimensions}}.}

\item{reorient}{Should image be reoriented if a filename?}
}
\value{
Object of class \code{nifti}, with binary values
}
\description{
Make a mask of an image that has all irrelevant
values
}
\note{
\code{empty_dim_mask} is a shorthand for \code{emptyImageDimensionsMask}
with all the same arguments.qQac vgfrew
}
\seealso{
\code{\link{getEmptyImageDimensions}}
}
