% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_img.R
\name{finite_img-methods}
\alias{finite_img-methods}
\alias{finite_img,antsImage-method}
\title{Finite Image}
\usage{
\S4method{finite_img}{antsImage}(img, replace = 0)
}
\arguments{
\item{img}{character path of image or 
an object of class \code{nifti}, or list of images}

\item{replace}{Value to replace non-finite values to}
}
\value{
image object
}
\description{
Simple wrapper for setting non-finite values to zero
}
