% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flip_img.R
\name{flip_img}
\alias{flip_img}
\title{Flip NifTI Image}
\usage{
flip_img(img, x = FALSE, y = FALSE, z = FALSE, ...)
}
\arguments{
\item{img}{nifti object or character filename}

\item{x}{(logical) Flip x direction}

\item{y}{(logical) Flip y direction}

\item{z}{(logical) Flip z direction}

\item{...}{Arguments passed to \code{\link{check_nifti}}}
}
\value{
Object of class nifti
}
\description{
This image will flip x, y, or z direction
}

