% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslbet_robust.R
\name{fslbet_robust}
\alias{fslbet_robust}
\title{Robust Skull Stripping with COG estimation and Bias Correction}
\usage{
fslbet_robust(
  img,
  outfile = NULL,
  retimg = TRUE,
  correct = TRUE,
  correction = "N4",
  recog = TRUE,
  reorient = FALSE,
  bet.opts = "",
  nvoxels = 0,
  swapdim = TRUE,
  remove.neck = TRUE,
  remover = c("remove_neck", "double_remove_neck"),
  robust.mask = FALSE,
  rbm.voxels = 7,
  template.file = file.path(fsldir(), "data/standard", "MNI152_T1_1mm_brain.nii.gz"),
  template.mask = file.path(fsldir(), "data/standard",
    "MNI152_T1_1mm_brain_mask.nii.gz"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{img}{(character) File to be skull stripped or object of class
nifti}

\item{outfile}{(character) output filename}

\item{retimg}{(logical) return image of class nifti}

\item{correct}{Perform bias field correction}

\item{correction}{(character) N3 or N4 correction?}

\item{recog}{Rerun bet with a new center of gravity (COG) estimate}

\item{reorient}{(logical) If retimg, should file be reoriented when read in?
Passed to \code{\link{readnii}}.}

\item{bet.opts}{Options for \code{\link{fslbet}}}

\item{nvoxels}{Number of voxels to dilate/erode.  See \code{\link{fslfill2}}.
If \code{nvoxels = 0}, then no smoothing is done.}

\item{swapdim}{Use \code{\link{fslswapdim}} to reorient image}

\item{remove.neck}{Run \code{\link{remove_neck}} to register the template to a 
thresholded image to remove neck slices.}

\item{remover}{if \code{remove.neck = TRUE}, then which function
would you like to use to remove the neck}

\item{robust.mask}{Run \code{\link{robust_brain_mask}} to register the template to a 
thresholded image and inflate for}

\item{rbm.voxels}{Number of voxels to inflate mask for \code{\link{robust_brain_mask}}}

\item{template.file}{Template to warp to original image space, passed to 
\code{\link{remove_neck}}}

\item{template.mask}{Mask of template to use as rough brain mask, passed 
to \code{\link{remove_neck}}}

\item{verbose}{(logical) Should diagnostic output be printed?}

\item{...}{additional arguments passed to \code{\link{robust_brain_mask}} or 
\code{\link{remove_neck}}.}
}
\value{
Skull-stripped \code{nifti} object
}
\description{
Skull Stripping (using FSL's BET) a file using \code{fslr}
functions and robustified by registration and neck removal
}
\note{
This function first thresholds an image, runs a rigid registration
(default in \code{\link{remove_neck}}) to drop any slices below the transformed
skull stripped template to remove neck slices.  The neck-removed image is 
then skull stripped using defaults in \code{\link{fslbet}}.  A new 
center of gravity is estiamted using \code{\link{cog}}, then the image is
skull stripped again using the new cog. After the skull stripped mask is 
created, the image is dilated and eroded using \code{\link{fslfill2}} to
fill holes using a box kernel with the number of voxels \code{nvoxels} in 
all 3 directions.
}
\examples{
\dontrun{
  stubs = c("T1Strip.nii.gz",
  "T2Strip.nii.gz")
  img_files = system.file(stubs,
  package="WhiteStripe")
  
  if (all(file.exists(img_files)){
   fslbet_robust(system.file('T1Strip.nii.gz', package="WhiteStripe"))
  }
}
}
