% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_xyz2.R
\name{group_xyz2}
\alias{group_xyz2}
\title{Center of Gravity for Multiple Components/Areas}
\usage{
group_xyz2(img, k = 1)
}
\arguments{
\item{img}{Object of class \code{\link{nifti}} or array}

\item{k}{Minimum number of voxels for a cluster/component.
See \code{\link{label_mask}}}
}
\value{
Matrix of 3 columns of dimension indices.
}
\description{
Find center of gravity for each compoent of an image
}

