% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_func.R
\name{image_func}
\alias{image_func}
\title{Generic Image Function}
\usage{
image_func(imgs, func = c("mean", "median", "sd", "var", "mad", "sum", "prod",
  "z"), finite = TRUE, ...)
}
\arguments{
\item{imgs}{Character vector, list of characters, or object of class 
\code{nifti}}

\item{func}{Function to perform voxel-wise on list of images}

\item{finite}{Should non-finite values be removed?}

\item{...}{Addictional arguments to pass to \code{func}}
}
\value{
Object of class \code{nifti}
}
\description{
Creates output from a list of images, performing voxel-wise
operations
}

