% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_similarity.R
\name{image_similarity}
\alias{image_similarity}
\title{Compute Image Similarity Metrics}
\usage{
image_similarity(fixed, moving, fixed_mask = NULL, moving_mask = NULL, ...)
}
\arguments{
\item{fixed}{Fixed image (not moving)}

\item{moving}{Moving image (moved into fixed space)}

\item{fixed_mask}{Mask for fixed image (optional)}

\item{moving_mask}{Mask for moving image (optional)}

\item{...}{Additional arguments to pass to
\code{\link{imageSimilarity}}}
}
\value{
A numeric value
}
\description{
Compute Image Similarity Metrics
}
\examples{
library(ANTsRCore)
library(extrantsr)
x =  getANTsRData( 'r16' )
y =  getANTsRData( 'r30' )
image_similarity(x,y, type="MeanSquares")
image_similarity(y,x, type="MeanSquares")

image_similarity(x,y, type="Correlation")
image_similarity(y,x, type="Correlation")

image_similarity(x,y, type="MattesMutualInformation")
image_similarity(y,x, type="MattesMutualInformation")

image_similarity(x,y, type="ANTSNeighborhoodCorrelation")
image_similarity(y,x, type="ANTSNeighborhoodCorrelation")
image_similarity(x,y, type="JointHistogramMutualInformation")
image_similarity(y,x, type="JointHistogramMutualInformation")
image_similarity(x,y, type="Demons")
image_similarity(y,x, type="Demons")
}
