% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_cov_matrix.R
\name{img_cov_matrix}
\alias{img_cov_matrix}
\title{Neighborhood Covariance Matrix}
\usage{
img_cov_matrix(imgs, mask = NULL, radius = rep(1, 3),
  method = c("pearson", "spearman"), center = TRUE, scale = FALSE,
  verbose = TRUE)
}
\arguments{
\item{imgs}{list of images of class \code{nifti}, \code{antsImage}, \code{character}}

\item{mask}{Binary image of class \code{nifti}, 
\code{antsImage}, \code{character}.  Only neighborhoods inside the mask 
will be taken}

\item{radius}{vector of length 3 for number of voxels to go in each direction.
Default is 27 neighbors (including voxel at center).  
Passed to \code{\link{getNeighborhoodInMask}}.}

\item{method}{Type of correlation. If pearson, scaling done on original data.
If spearman, the ranks. If you want a covariance, you simply need to do 
\code{scale = FALSE}}

\item{center}{Should the data be centered}

\item{scale}{Should the data be scaled}

\item{verbose}{print diagnostic messages}
}
\value{
List of matrix of covariances for each voxel and the order the columns
are in.

#' @export
}
\description{
Calculates the neighborhood of multiple images, and then
does the covariance of each neighbor with itself in the other imaging modality
}
\note{
Divisor is \code{n - 1}.
}
