% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_density_entropy.R
\name{img_density_entropy}
\alias{img_density_entropy}
\title{Image Entropy using density}
\usage{
img_density_entropy(img, mask = NULL, base = exp(1), ...)
}
\arguments{
\item{img}{object of class \code{nifti} or character}

\item{mask}{binary object of class \code{nifti} or character}

\item{base}{Base to take the log for the entropy.  Default is e}

\item{...}{Additional arguments to pass to \code{\link{density}}.}
}
\description{
Get the entropy for an entire image using density binning
}
\examples{
library(oro.nifti)
arr = array(rnorm(100^3), dim = rep(100, 3))
img = oro.nifti::nifti(arr)
img_density_entropy(img)
}
