% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_hist_entropy.R
\name{img_hist_entropy}
\alias{img_hist_entropy}
\title{Image Entropy using histograms}
\usage{
img_hist_entropy(img, mask = NULL, breaks = 2000, base = exp(1), ...)
}
\arguments{
\item{img}{object of class \code{nifti} or character}

\item{mask}{binary object of class \code{nifti} or character}

\item{breaks}{Number of breaks to pass to \code{\link{hist}}}

\item{base}{Base to take the log for the entropy.  Default is e}

\item{...}{Additional arguments to pass to \code{hist}.  Plot has already been
set to \code{FALSE}}
}
\description{
Get the entropy for an entire image using histogram binning
}
\examples{
library(oro.nifti)
arr = array(rnorm(100^3), dim = rep(100, 3))
img = oro.nifti::nifti(arr)
img_hist_entropy(img)
}
