% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_ts_to_list.R
\name{img_list_to_ts}
\alias{img_list_to_ts}
\title{Image List to Time Series}
\usage{
img_list_to_ts(imgs, copy_nifti = TRUE, warn = TRUE)
}
\arguments{
\item{imgs}{object of class \code{\link{list}}, each with 3 dimensions,}

\item{copy_nifti}{Should a \code{nifti} object be returned (\code{TRUE}) or 
a simply array (\code{FALSE}).  Should only be used for slight speed up when
array is adequate}

\item{warn}{Should a warning be printed if object is not class
\code{\link{nifti}}}
}
\value{
Object of class \code{\link{nifti}}
}
\description{
Turns a a list of 3D images into a 4D time series image
}
\note{
If the object is not of class \code{\link{list}},
then the object is returned
}

