% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_split.R
\name{img_split}
\alias{img_split}
\title{Split an Image into Multiple Sub-Images}
\usage{
img_split(img, output_basename = NULL)
}
\arguments{
\item{img}{Object of class \code{\link{antsImage}}}

\item{output_basename}{}
}
\value{
Names of output files
}
\description{
Split an image on a certain dimension
}

