% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_ts_to_list.R
\name{img_ts_to_df}
\alias{img_ts_to_df}
\title{Image Time Series to Data.frame}
\usage{
img_ts_to_df(imgs, warn = FALSE)
}
\arguments{
\item{imgs}{object of class \code{\link{nifti}} with 4 dimensions, 
aka a 4D time series}

\item{warn}{Should a warning be printed if object is not class
\code{\link{nifti}} (e.g. a list instead)}
}
\value{
Matrix of values
}
\description{
Turns a 4D time series image to a Data.frame
}
