% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpaint.R
\name{inpaint}
\alias{inpaint}
\title{Inpainting of images}
\usage{
inpaint(img, input_mask, fill_mask, func = c("mean", "median"), ...)
}
\arguments{
\item{img}{image of class \code{nifti} or character}

\item{input_mask}{mask to be painted in.}

\item{fill_mask}{image mask to take the function (e.g. mean/median) of the image}

\item{func}{Function to apply on the values.  \code{do.call} will be called on them}

\item{...}{Arguments to be passed to \code{func}.}
}
\value{
List of output image (\code{outimg}) and 
statistic filled in (\code{statistic})
}
\description{
Paints in the image with the mean/median from 
\code{fill_mask}
}

