% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invwarp.R
\name{invwarp}
\alias{invwarp}
\title{Inverse Warp from antsRegistration}
\usage{
invwarp(dimension, fixed, output, moving, transformlist)
}
\arguments{
\item{dimension}{Number of dimensions of the input image}

\item{fixed}{fixed image to which we register the moving image.}

\item{output}{output filename}

\item{moving}{moving image to be mapped to fixed space.}

\item{transformlist}{list of transforms usually from \code{antsRegistration}}
}
\value{
Result of \code{system} command
}
\description{
A simplified (or full) interface to 
WarpImageMultiTransform.
}

