% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobian.R
\name{jacobian_image}
\alias{jacobian_image}
\alias{log_jacobian_image}
\alias{geom_jacobian_image}
\alias{log_geom_jacobian_image}
\title{Image of the Determinant of the Jacobian}
\usage{
jacobian_image(fixed, transformlist = NULL, moving = NULL,
  typeofTransform = "SyN", ...)

log_jacobian_image(fixed, transformlist = NULL, moving = NULL,
  typeofTransform = "SyN", ...)

geom_jacobian_image(fixed, transformlist = NULL, moving = NULL,
  typeofTransform = "SyN", ...)

log_geom_jacobian_image(fixed, transformlist = NULL, moving = NULL,
  typeofTransform = "SyN", ...)
}
\arguments{
\item{fixed}{Fixed image in a registration}

\item{transformlist}{List of transformation of moving image to
fixed image.  Only one of \code{transformlist} or \code{moving} 
should be specified.}

\item{moving}{Image to transform into \code{fixed} space, if not 
done previously. Only one of \code{transformlist} or \code{moving} 
should be specified.}

\item{typeofTransform}{If registration not done previously,
then transform to use for transformation.  Should be non-linear.}

\item{...}{additional arguments to pass to \code{\link{registration}}}
}
\value{
An image of the determinant of the Jacobian, the
transform list, and the fixed image.
}
\description{
Image of the Determinant of the Jacobian
}
