% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_bwlabel.R
\name{label_clusters}
\alias{label_clusters}
\title{Label Clusters}
\usage{
label_clusters(img, ..., retfile = FALSE)
}
\arguments{
\item{img}{input image}

\item{...}{additional arguments passed to \code{\link{labelClusters}}}

\item{retfile}{logical to indicate if an \code{antsImage} should be returned
\code{TRUE} (useful for chaining) or a \code{nifti} image}
}
\value{
A \code{antsImage} or a \code{nifti} image, depending on
\code{retfile}
}
\description{
Label Clusters
}
