% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_xyz2.R
\name{label_mask}
\alias{label_mask}
\title{Label Connected Components}
\usage{
label_mask(img, k = 1)
}
\arguments{
\item{img}{Object of class \code{\link{nifti}} or array}

\item{k}{Minimum number of voxels for a cluster/component.}
}
\value{
Object of class \code{\link{nifti}}
}
\description{
Label a binary image with connected components
}

