% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largest_component.R
\name{largest_component}
\alias{largest_component}
\title{Get Largest Component}
\usage{
largest_component(img, ..., retfile = FALSE)
}
\arguments{
\item{img}{Object of class \code{nifti}}

\item{...}{Additional arguments passed to \code{\link{iMath}}}

\item{retfile}{logical to indicate if an \code{antsImage} should be returned
(useful for chaining)}
}
\value{
Object of class \code{nifti}
}
\description{
Wrapper for \code{oMath("GetLargestComponent")}
}
