% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_quantile_img.R
\name{local_quantile_img}
\alias{local_quantile_img}
\title{Quantile Transform Voxel based on Neighborhood Distribution}
\usage{
local_quantile_img(img, mask = NULL, radius = rep(1, 3),
  verbose = TRUE)
}
\arguments{
\item{img}{Object of class \code{\link{nifti}}, character or \code{antsImage}}

\item{mask}{Binary image of class \code{nifti}, 
\code{antsImage}, \code{character}.  Only neighborhoods inside the mask 
will be taken}

\item{radius}{vector of length 3 for number of voxels to go in each direction.
Default is 27 neighbors (including voxel at center).  
Passed to \code{\link{getNeighborhoodInMask}}.}

\item{verbose}{Print diagnostic messages}
}
\value{
An object of class \code{nifti}
}
\description{
Quantile Transform Voxel based on Neighborhood Distribution
}
\examples{
nim = oro.nifti::nifti(array(rnorm(10^3), dim = rep(10, 3)))
mask = nim > -1
res = local_quantile_img(nim, mask = mask)

}
