% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makepng}
\alias{makepng}
\title{Non-exported function to make png}
\usage{
makepng(addstub, ...)
}
\arguments{
\item{addstub}{extension to add to outfile}

\item{...}{arguments passed to \code{\link{png}}}
}
\value{
Alpha level depending on device
}
\description{
Make my png
}

