% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/malf.R
\name{malf}
\alias{malf}
\title{Multi-Atlas Label Fusion}
\usage{
malf(infile, template.images, template.structs, keep_images = TRUE,
  outfiles = NULL, outfile = NULL, retimg = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{infile}{Input Image file}

\item{template.structs}{Template gold standards to apply
registration into \code{infile} space}

\item{keep_images}{Keep the \code{template.structs} in
\code{infile} space}

\item{outfiles}{Output filenames for  \code{template.structs} in
\code{infile} space}

\item{outfile}{Fused output filename}

\item{retimg}{Return Image to user using \code{\link{readNIfTI}}}

\item{verbose}{Print diagnostic output}

\item{...}{Arguments to be passed to \code{\link{ants_regwrite}}}

\item{template.image}{Template Images to register
to \code{infile}}
}
\value{
The output filename or the nifti image
}
\description{
Takes in an input file and template images with
a set of template structures and creates a label fusion
}

