% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malf_registration.R
\name{malf_registration}
\alias{malf_registration}
\title{Multi-Atlas Label Fusion Registration}
\usage{
malf_registration(infile, template.images, template.structs,
  typeofTransform = "SyN", interpolator = "NearestNeighbor",
  keep_images = TRUE, outfiles = NULL, outprefix = NULL, verbose = TRUE,
  ...)
}
\arguments{
\item{infile}{Input Image file}

\item{template.images}{Template Images to register 
to \code{infile}}

\item{template.structs}{Template gold standards to apply 
registration into \code{infile} space}

\item{typeofTransform}{type of transformed used, passed to 
\code{\link{antsRegistration}}}

\item{interpolator}{interpolation done for 
\code{\link{antsApplyTransforms}}}

\item{keep_images}{Keep the \code{template.structs} in 
\code{infile} space}

\item{outfiles}{Output filenames for  \code{template.structs} in 
\code{infile} space}

\item{outprefix}{passed to \code{\link{registration}} if 
\code{keep_regs = TRUE}}

\item{verbose}{Print diagnostic output}

\item{...}{Arguments to be passed to \code{\link{registration}}}
}
\value{
List of registrations and
output files
}
\description{
Takes in an input file and template images with 
a set of template structures and registers them
}

