% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskEmptyImageDimensions.R
\docType{methods}
\name{maskEmptyImageDimensions,antsImage-method}
\alias{maskEmptyImageDimensions,antsImage-method}
\alias{maskEmptyImageDimensions,nifti-method}
\title{Apply Masking from Empty Image Dimensions}
\usage{
\S4method{maskEmptyImageDimensions}{antsImage}(img, inds,
  reorient = FALSE, mask.value = 0, ...)
}
\arguments{
\item{img}{image, nifti object, or array}

\item{inds}{indices of subset from \code{\link{getEmptyImageDimensions}} or
\code{\link{dropEmptyImageDimensions}}.}

\item{reorient}{Should image be reoriented if a filename?}

\item{mask.value}{Value to replace voxels outside the mask.}

\item{...}{not used}
}
\value{
Object of class \code{nifti} or \code{array} if \code{nifti}
is not supplied
}
\description{
Simple wrapper for replacing indices with a value
}
\note{
\code{apply_empty_dim} is a shorthand for 
\code{maskEmptyImageDimensions} with all the same arguments.
}
\seealso{
\code{\link{getEmptyImageDimensions}}, 
\code{\link{dropEmptyImageDimensions}}
}
