% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_reg.R
\name{multi_reg}
\alias{multi_reg}
\title{Multi-Registration}
\usage{
multi_reg(infiles, template.file, outfiles = NULL, typeofTransform = "SyN",
  interpolator = "Linear", outprefix = NULL, retimg = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{outfiles}{Output filenames for  \code{template.structs} in 
\code{infile} space}

\item{typeofTransform}{type of transformed used, passed to 
\code{\link{antsRegistration}}}

\item{interpolator}{interpolation done for 
\code{\link{antsApplyTransforms}}}

\item{retimg}{Return image to user using \code{\link{readNIfTI}}}

\item{verbose}{Print diagnostic output}

\item{...}{Arguments to be passed to \code{\link{ants_regwrite}}}

\item{infile}{Input image file}

\item{template.images}{Template images to register 
to \code{infile}}

\item{template.structs}{Template gold standards to apply 
registration into \code{infile} space}

\item{keep_images}{Keep the \code{template.structs} in 
\code{infile} space}

\item{outfile}{Fused output filename}
}
\value{
Output list of registered images and transformations
}
\description{
Takes a list of images and registers them to a template, 
keeping the registration information in a list.
}

