% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_rereg.R
\name{multi_rereg}
\alias{multi_rereg}
\title{Two-Stage Multi-Registration of images}
\usage{
multi_rereg(infiles, func = "mean", ...)
}
\arguments{
\item{infiles}{Input image files}

\item{func}{Function to be passed to}

\item{...}{Arguments to be passed to \code{\link{multi_reg}}.}
}
\value{
Output list of registered images to statistic image 
and transformations
}
\description{
Takes a list of images and registers them to a template, 
takes a statistic image, and registers the list of images back to this 
statistic image.
}
\note{
The second stage will use the same arguments passed in \code{...}, but
the \code{template.file} will be the statistic image
}

