% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_information.R
\name{mutual_information}
\alias{mutual_information}
\title{Calculates Mutual Information on nifti objects}
\usage{
mutual_information(image1, image2, ...)
}
\arguments{
\item{image1}{Object of class \code{nifti} or character filename}

\item{image2}{Object of class \code{nifti} or character filename}

\item{...}{Additional arguments passed to 
\code{\link{antsImageMutualInformation}}}
}
\value{
Numeric
}
\description{
Images are converted \code{nifti} objects
(from the oro.nifti package), an operation is performed using 
\code{\link{antsImageMutualInformation}}
}
\examples{

fi = getANTsRData("r16")
mi = getANTsRData("r64")
mival = mutual_information(fi, mi)
}
