% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n4_4d.R
\name{n4_4d}
\alias{n4_4d}
\title{Four-Dimensional Bias-Field Correction}
\usage{
n4_4d(img, verbose = TRUE)
}
\arguments{
\item{img}{Image for N4 to be run, an antsImage}

\item{verbose}{Print diagnostic messages}
}
\value{
An object of class \code{antsImage}
}
\description{
Four-Dimensional Bias-Field Correction
}
