% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.R
\name{neighborhood}
\alias{neighborhood}
\title{Get Neighborhood of Voxels from an Image}
\usage{
neighborhood(img, mask = NULL, radius = rep(1, 3), get.gradient = TRUE,
  verbose = TRUE, run_gc = TRUE, ...)
}
\arguments{
\item{img}{Object of class \code{\link{nifti}}, character or \code{antsImage}}

\item{mask}{Binary image of class \code{nifti}, 
\code{antsImage}, \code{character}.  Only neighborhoods inside the mask 
will be taken}

\item{radius}{vector of length 3 for number of voxels to go in each direction.
Default is 27 neighbors (including voxel at center).  
Passed to \code{\link{getNeighborhoodInMask}}.}

\item{get.gradient}{logical indicating if a matrix of gradients (at the center voxel) 
should be returned in addition to the value matrix}

\item{verbose}{Print diagnostic messages}

\item{run_gc}{Experimental - trying to run with gc(); for memory cleanup.}

\item{...}{arguments other than \code{spatial.info} passed to 
\code{\link{getNeighborhoodInMask}}}
}
\value{
List similar to the output of 
\code{\link{getNeighborhoodInMask}}
}
\description{
This function wraps the \code{\link{getNeighborhoodInMask}} to
be more friendly for R and \code{nifti} objects
}
