% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis.R
\name{oasis}
\alias{oasis}
\title{OASIS Processing Pipeline}
\usage{
oasis(filename, skull_strip = TRUE, skull_stripfile = NULL,
  correct = TRUE, normalize = TRUE, normalize_file = NULL,
  retimg = TRUE, outfile = NULL, template.file = file.path(fsldir(),
  "data", "standard", "MNI152_T1_1mm_brain.nii.gz"),
  interpolator = "LanczosWindowedSinc", other.files = NULL,
  other.outfiles = NULL, typeofTransform = "Rigid",
  remove.warp = TRUE, outprefix = NULL, bet.opts = "-B -f 0.1 -v",
  betcmd = "bet", reproducible = TRUE, seed = 1, ...)
}
\arguments{
\item{filename}{filename of T1 image}

\item{skull_strip}{do skull stripping with FSL BET}

\item{skull_stripfile}{Output skull strip filename}

\item{correct}{do N3 Bias correction}

\item{normalize}{Normalize data using \code{\link{whitestripe}}}

\item{normalize_file}{\code{\link{whitestripe}} image mask}

\item{retimg}{return a nifti object from function}

\item{outfile}{output filename should have .nii or .nii.gz 
extension}

\item{template.file}{Filename of template to warp to}

\item{interpolator}{interpolation done for 
\code{\link{antsApplyTransforms}}}

\item{other.files}{Filenames of other iamges to be 
transformed with the T1}

\item{other.outfiles}{Output filenames of \code{other.files} to 
be written}

\item{typeofTransform}{type of transformed used, passed to 
\code{\link{antsRegistration}}}

\item{remove.warp}{(logical) Should warping images be deleted?}

\item{outprefix}{Character path of where the warp files should be stored.
Required if \code{remove.warp = FALSE}}

\item{bet.opts}{Options passed to \code{\link{fslbet}}}

\item{betcmd}{BET command used, passed to \code{\link{fslbet}}}

\item{reproducible}{Sets the seed and 
\code{Sys.setenv(ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS = 1)}.
 See
\url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
for discussion.}

\item{seed}{will execute 
\code{Sys.setenv(ANTS_RANDOM_SEED = seed)} before
running to attempt a more reproducible result.   If \code{NULL}, will not set anything, 
but \code{reproducible} must be \code{FALSE}.}

\item{...}{arguments to \code{\link{whitestripe}}}
}
\value{
NULL or object of class nifti for transformed T1 image
}
\description{
This function performs registration to a T1 template
using ANTsR transformation
}
