% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oMask.R
\name{oMask}
\alias{oMask}
\alias{get_mask}
\title{Get Mask from nifti object}
\usage{
oMask(img, ...)

get_mask(img, ...)
}
\arguments{
\item{img}{Object of class nifti or character filename}

\item{...}{arguments to \code{\link{getMask}}}
}
\value{
Object of class \code{nifti}
}
\description{
Wraps \code{getMask} from \code{ANTsR} for 
nifti objects
}
