% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oMath.R
\name{oMath}
\alias{oMath}
\title{Perform iMath on nifti objects}
\usage{
oMath(img, ...)
}
\arguments{
\item{img}{Object of class \code{nifti}}

\item{...}{Additional arguments passed to \code{\link{iMath}}}
}
\value{
Object of class \code{nifti}
}
\description{
Images are converted \code{nifti} objects
(from the oro.nifti package), an operation is performed using
\code{\link{iMath}}, and the result is a \code{nifti} object.
}

