% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho2.R
\docType{methods}
\name{ortho2,antsImage-method}
\alias{ortho2,antsImage,ANY,ANY-method}
\alias{ortho2,antsImage-method}
\title{Orthographic Display, added options}
\usage{
\S4method{ortho2}{antsImage}(x, y = NULL, xyz = NULL, w = 1,
  col = gray(0:64/64), col.y = hotmetal(), zlim = NULL, zlim.y = NULL,
  NA.x = FALSE, NA.y = TRUE, crosshairs = TRUE, col.crosshairs = "red",
  xlab = "", ylab = "", axes = FALSE, oma = c(0, 0, 0, ifelse(ycolorbar,
  5, 0)), mar = rep(0, 4), bg = "black", text = NULL,
  text.color = "white", text.cex = 2, text.x = 32, text.y = 32,
  add.orient = TRUE, mfrow = c(2, 2), ybreaks = NULL, breaks = NULL,
  addlegend = FALSE, leg.x = 32, leg.y = 32, legend, leg.col,
  leg.title = NULL, leg.cex, window = NULL, ycolorbar = FALSE,
  clabels = TRUE, add = TRUE, pdim = NULL, ...)
}
\arguments{
\item{x}{is an object of class \code{antsImage} or similar.}

\item{y}{is an object of class \code{antsImage} or similar for the overlay.}

\item{pdim}{Pixel dimensions if passing in arrays.}

\item{...}{other arguments to the image function may be provided here.}
}
\description{
Wrapper for of \code{fslr}'s \code{\link{ortho2}} function 
to work with antsImages
}
\note{
This will NOT do \code{\link{ants2oro}} to plot the image.  
If that is desired, users will have to do that before calling 
\code{ortho2}.
}

