% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho2.R
\name{ortho2}
\alias{ortho2}
\title{Orthographic Display, added options}
\usage{
ortho2(x, y = NULL, pdim = NULL, ...)
}
\arguments{
\item{x}{is an object of class \code{antsImage} or similar.}

\item{y}{is an object of class \code{antsImage} or similar for the overlay.}

\item{pdim}{Pixel dimensions if passing in arrays.}

\item{...}{other arguments to the image function may be provided here.}
}
\description{
Wrapper for of \code{fslr}'s \code{\link{ortho2}} function 
to work with antsImages
}
\note{
This will NOT do \code{\link{ants2oro}} to plot the image.  
If that is desired, users will have to do that before calling 
\code{ortho2}.
}

