% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/otropos.R
\name{otropos}
\alias{otropos}
\title{Run Atropos for nifti objects}
\usage{
otropos(a, x, ..., make_mask = TRUE, reset_origin = TRUE)
}
\arguments{
\item{a}{Images to segment.  Can be a \code{nifti} object,
\code{antsImage} object, \code{character} vector, or list of these}

\item{x}{Mask image.  If none is supplied and \code{make_mask = TRUE},
then mask is created by first}

\item{...}{Additional arguments to be passed to \code{\link{atropos}}}

\item{make_mask}{Logical indicating if mask should be created if
\code{x} is missing}

\item{reset_orogin}{Should \code{\link{antsCopyOrigin}} be done
on the mask?}
}
\value{
Result of \code{\link{atropos}}, but with nifti objects.
}
\description{
Runs Atropos (a finite mixture model) but with more
robust inputs
}

