% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perona_malik.R
\name{perona_malik}
\alias{perona_malik}
\title{Perona-Malik Anisotropic Filter}
\usage{
perona_malik(img, n_iter, conductance, ..., retfile)
}
\arguments{
\item{img}{Object of class \code{nifti}}

\item{n_iter}{number of iterations}

\item{conductance}{From ITK: The conductance parameter controls the 
sensitivity of the conductance term in the basic anisotropic diffusion equation}

\item{...}{Additional arguments passed to \code{\link{iMath("PeronaMalik")}}}

\item{retfile}{logical to indicate if an \code{antsImage} should be returned
(useful for chaining)}
}
\value{
Object of class \code{nifti}
}
\description{
Wrapper for \code{oMath("PeronaMalik")}
}
