% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sublime.R
\name{preprocess_mri_across}
\alias{preprocess_mri_across}
\title{Preprocess MRI acros visits}
\usage{
preprocess_mri_across(baseline_files, followup_files, baseline_outfiles,
  followup_outfiles, retimg = FALSE, maskfile = NULL, skull_strip = FALSE,
  bet.opts = "-B -f 0.1 -v", betcmd = "bet", within.transform = "Rigid",
  within.interpolator = "LanczosWindowedSinc",
  across.transform = within.transform,
  across.interpolator = within.interpolator, verbose = TRUE, ...)
}
\arguments{
\item{baseline_files}{filename of baseline images}

\item{followup_files}{filename of followup images}

\item{baseline_outfiles}{output filenames for baseline images}

\item{followup_outfiles}{output filenames for followup images}

\item{retimg}{}

\item{maskfile}{}

\item{skull_strip}{do Skull stripping with FSL BET}

\item{bet.opts}{}

\item{betcmd}{}

\item{within.transform}{Transformation for within-visit registration}

\item{within.interpolator}{Interpolator for within-visit registration}

\item{across.transform}{Transformation for across-visit registration}

\item{across.interpolator}{Interpolator for across-visit registration}

\item{verbose}{print diagnostic outputs}

\item{...}{arguments to \code{\link{preprocess_mri_within}} and \code{\link{ants_regwrite}}}
}
\value{
NULL or list of objects of class nifti for transformed images
}
\description{
This function performs preprocessing and registration
across visits and within visits
using ANTsR and SyN transformation
}

