% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_mri_within.R
\name{preprocess_mri_within}
\alias{preprocess_mri_within}
\title{OASIS Processing Pipeline}
\usage{
preprocess_mri_within(files, outfiles = NULL, correct = TRUE,
  correction = "N3", shrinkfactor = "4", retimg = FALSE,
  reorient = FALSE, typeofTransform = "Rigid",
  interpolator = "LanczosWindowedSinc", skull_strip = FALSE,
  bet.opts = "-B -f 0.1 -v", betcmd = "bet", maskfile = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{files}{filenames (or nifti objects) of images to be processed.
Will register to the first scan}

\item{outfiles}{(character) name of output files, with extension}

\item{correct}{do Bias field correction}

\item{correction}{(character) N3 or N4 correction?}

\item{shrinkfactor}{Shrink factor passed to 
\code{\link{n3BiasFieldCorrection}}}

\item{retimg}{(logical) return list of images of class nifti}

\item{reorient}{(logical) If retimg, should file be reoriented 
when read in?
Passed to \code{\link{readNIfTI}}.}

\item{typeofTransform}{type of transformed used, passed to 
\code{\link{antsRegistration}}}

\item{interpolator}{Interpolation to be performed, passed to
\code{\link{antsRegistration}}}

\item{skull_strip}{do Skull stripping with FSL BET}

\item{bet.opts}{Options to pass to \code{\link{fslbet}}}

\item{betcmd}{Command to pass to \code{\link{fslbet}}}

\item{maskfile}{Filename (or nifti object) of mask for image to be
registered to}

\item{verbose}{Diagnostic messages}

\item{...}{arguments to \code{\link{bias_correct}} or 
\code{\link{within_visit_registration}}}
}
\value{
List of outfiles, maskfile, and output from \code{\link{registration}}.
}
\description{
This function performs registration to a T1 template
using ANTsR and SyN transformation
}

