% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess_mri_across.R
\name{process_filenames}
\alias{process_filenames}
\title{Process filenames to ensure nifti and consistency}
\usage{
process_filenames(files, outfiles, force_nii = TRUE, copyfiles = TRUE)
}
\arguments{
\item{files}{input filenames}

\item{outfiles}{output filenames}

\item{force_nii}{Force output filenames to have .nii}

\item{copyfiles}{copy files to outfiles}
}
\value{
List of 2 elements: files of the filenames and outfiles
of output filenames
}
\description{
Takes input and output filename, ensures the same
file extension, copies files, and checks all are character
}

