% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transformlist.R
\name{read_transformlist}
\alias{read_transformlist}
\title{Read Transform List from ANTsR}
\usage{
read_transformlist(transformlist)
}
\arguments{
\item{transformlist}{Character vector of transform list}
}
\value{
List of transforms
}
\description{
Transform a transformlist from ANTsR into an object in R
}

