% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_flip.R
\name{reg_flip}
\alias{reg_flip}
\title{Flip Image with Registration}
\usage{
reg_flip(t1, register = TRUE, native = TRUE,
  template.file = file.path(fsldir(), "data", "standard",
  paste0("MNI152_T1_1mm", ifelse(is.null(mask), "", "_brain"), ".nii.gz")),
  typeofTransform = c("Rigid", "Affine"),
  interpolator = "LanczosWindowedSinc", t1.outfile = NULL,
  other.files = NULL, other.outfiles = NULL, mask = NULL,
  mask.outfile = NULL, verbose = TRUE, a = "-x", b = "y", c = "z",
  ...)
}
\arguments{
\item{t1}{filename (or nifti objects) of T1 image}

\item{register}{Register image to template file}

\item{native}{If images are registered, should the native space
normalized image be returned?}

\item{template.file}{Filename of template to warp to}

\item{typeofTransform}{type of transformed used, passed to 
\code{\link{antsRegistration}}}

\item{interpolator}{Interpolation to be performed, passed to
\code{\link{antsRegistration}}}

\item{t1.outfile}{Output filename of normalized T1 image}

\item{other.files}{Character filenames or list of nifti objects to 
normalize.  In the same space as T1.}

\item{other.outfiles}{Character filenames for output 
normalized files.}

\item{mask}{File or nifti image of brain mask}

\item{mask.outfile}{Character filename for output 
brain mask.}

\item{verbose}{Print Diagnostic Messages}

\item{a}{Option for x domain in \code{\link{fslswapdim}}}

\item{b}{Option for y domain in \code{\link{fslswapdim}}}

\item{c}{Option for z domain in \code{\link{fslswapdim}}}

\item{...}{arguments to \code{\link{fslswapdim}}}
}
\value{
List of nifti objects or character filenames
}
\description{
This function performs registration to a template
using ANTsR and will run \code{\link{fslswapdim}}
}
