% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reg_whitestripe.R
\name{reg_whitestripe}
\alias{reg_whitestripe}
\title{WhiteStripe Normalization}
\usage{
reg_whitestripe(t1 = NULL, t2 = NULL, register = TRUE, native = TRUE,
  template.file = file.path(fsldir(), "data", "standard",
  "MNI152_T1_1mm_brain.nii.gz"), typeofTransform = c("Rigid", "Affine"),
  interpolator = "LanczosWindowedSinc", type = c("T1", "T2", "hybrid"),
  t1.outfile = NULL, t2.outfile = NULL, other.files = NULL,
  other.outfiles = NULL, ...)
}
\arguments{
\item{typeofTransform}{type of transformed used, passed to
\code{\link{antsRegistration}}}

\item{interpolator}{Interpolation to be performed, passed to
\code{\link{antsRegistration}}}

\item{...}{arguments to \code{\link{bias_correct}} or
\code{\link{within_visit_registration}}}

\item{files}{filenames (or nifti objects) of images to be processed.
Will register to the first scan}

\item{outfiles}{(character) name of output files, with extension}

\item{n3correct}{do N3 Bias correction}

\item{correction}{(character) N3 or N4 correction?}

\item{shrinkfactor}{Shrink factor passed to
\code{\link{N3BiasFieldCorrection}}}

\item{retimg}{(logical) return list of images of class nifti}

\item{reorient}{(logical) If retimg, should file be
reoriented when read in?
Passed to \code{\link{readNIfTI}}.}

\item{skull_strip}{do Skull stripping with FSL BET}

\item{bet.opts}{Options to pass to \code{\link{fslbet}}}

\item{betcmd}{Command to pass to \code{\link{fslbet}}}

\item{maskfile}{Filename (or nifti object) of mask for image to be
registered to}

\item{verbose}{Diagnostic messages}
}
\value{
NULL or object of class nifti for transformed T1 image
}
\description{
This function performs registration to a T1 template
using ANTsR and SyN transformation
}

