% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_regwrite2.R
\name{registration}
\alias{registration}
\title{Registration Wrapper function}
\usage{
registration(filename, skull_strip = FALSE, correct = FALSE,
  correction = "N4", retimg = TRUE, outfile = NULL,
  template.file = file.path(fsldir(), "data", "standard",
  "MNI152_T1_1mm_brain.nii.gz"), interpolator = "Linear",
  other_interpolator = interpolator, other.files = NULL,
  other.outfiles = NULL, other.init = NULL,
  invert_interpolator = interpolator, invert.native.fname = NULL,
  invert.file = NULL, typeofTransform = "SyN", remove.warp = FALSE,
  outprefix = NULL, bet.opts = "-B -f 0.1 -v", betcmd = "bet",
  copy_origin = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{filename}{filename of image to be registered}

\item{skull_strip}{do skull stripping with FSL BET}

\item{correct}{do bias correction on the filename}

\item{correction}{N3 or N4 correction, see \code{\link{bias_correct}}}

\item{retimg}{return a nifti object from function}

\item{outfile}{output filename should have .nii or .nii.gz
extension}

\item{template.file}{Filename of image to warp to}

\item{interpolator}{interpolation done for
\code{\link{antsApplyTransforms}}}

\item{other_interpolator}{interpolation done for
\code{\link{antsApplyTransforms}} for \code{other.files}}

\item{other.files}{Filenames of other iamges to be
transformed with the T1}

\item{other.outfiles}{Output filenames of \code{other.files} to
be written}

\item{other.init}{Initial transformation lists (same length as \code{other.files})
to use in before the estimated transformation for one interpolation}

\item{invert_interpolator}{interpolation done for
\code{\link{antsApplyTransforms}} for \code{invert.file}}

\item{invert.native.fname}{filename of output native file, must have
same length as \code{invert.file} or be \code{NULL}.}

\item{invert.file}{Filename of image to invert to native space}

\item{typeofTransform}{type of transformed used, passed to
\code{\link{antsRegistration}}}

\item{remove.warp}{(logical) Should warping images be deleted?}

\item{outprefix}{Character path of where the warp files should be stored.
Required if \code{remove.warp = FALSE}}

\item{bet.opts}{Options passed to \code{\link{fslbet}}}

\item{betcmd}{BET command used, passed to \code{\link{fslbet}}}

\item{copy_origin}{Copy image origin from t1, using \code{\link{antsCopyOrigin}}}

\item{verbose}{Print diagnostic messages}

\item{...}{arguments to \code{\link{antsRegistration}}}
}
\value{
List of the output filenames and transformations
}
\description{
This function performs registration
using ANTsR, carries out the transformation
on other images, and can back-transform image
in registered space to the native space of the original image.  Returns the
transforms
}
