% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_neighborhood.R
\name{reorder_neigh_indices}
\alias{reorder_neigh_indices}
\title{Reorder Voxel Neighborhood}
\usage{
reorder_neigh_indices(indices, img.dim)
}
\arguments{
\item{indices}{Indices from \code{\link{getNeighborhoodInMask}}}

\item{img.dim}{Image dimension from which neighborhood was taken}
}
\value{
List of new indices, the ordering, and a binary array which has a 
1 where the indices are located
}
\description{
The output from \code{\link{getNeighborhoodInMask}} gets reordered
to match that which R uses for array to indices.
}
