% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducible_ants.R
\name{reproducible_ants}
\alias{reproducible_ants}
\title{Set reproducible seeds and threads for ANTsR execution}
\usage{
reproducible_ants(seed = 1)
}
\arguments{
\item{seed}{will execute 
\code{Sys.setenv(ANTS_RANDOM_SEED = seed)} before
running to attempt a more reproducible result.}
}
\value{
A list of `ANTS_RANDOM_SEED` and `ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS`
}
\description{
Set reproducible seeds and threads for ANTsR execution
}
\note{
See
\url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
for discussion.
}
\examples{
reproducible_ants(1234)
}
