% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_image.R
\docType{methods}
\name{resample_image-methods}
\alias{resample_image-methods}
\alias{resample_image}
\alias{resample_image,character-method}
\alias{resample_image,nifti-method}
\alias{resample_image,antsImage-method}
\alias{.resample_image}
\title{Resample an image}
\usage{
resample_image(img, parameters, parameter_type = c("mm", "voxels"),
  interpolator = c("nearestneighbor", "linear", "gaussian", "windowedsinc",
  "bspline"), ...)

\S4method{resample_image}{character}(img, parameters, parameter_type = c("mm",
  "voxels"), interpolator = c("nearestneighbor", "linear", "gaussian",
  "windowedsinc", "bspline"), ...)

\S4method{resample_image}{nifti}(img, parameters, parameter_type = c("mm",
  "voxels"), interpolator = c("nearestneighbor", "linear", "gaussian",
  "windowedsinc", "bspline"), ...)

\S4method{resample_image}{antsImage}(img, parameters, parameter_type = c("mm",
  "voxels"), interpolator = c("nearestneighbor", "linear", "gaussian",
  "windowedsinc", "bspline"), ...)

.resample_image(img, parameters, parameter_type = c("mm", "voxels"),
  interpolator = c("nearestneighbor", "linear", "gaussian", "windowedsinc",
  "bspline"))
}
\arguments{
\item{img}{character path of image or 
an object of class nifti, or antsImage}

\item{parameters}{parameters
to pass as \code{resampleParams} in
\code{\link{resampleImage}}}

\item{parameter_type}{Are the parameters in units of 
voxels or millimeters (mm)}

\item{interpolator}{what type of interpolation should be
done}

\item{...}{not currently used}
}
\value{
An \code{antsImage} or \code{\link{nifti}} depending on
input
}
\description{
Resamples an \code{antsImage} or 
\code{\link{nifti}}
object
}
\examples{
library(oro.nifti)
library(extrantsr)
n = 30
x = nifti(array(rnorm(n^3*10), dim = c(n, n, n)))
res = resample_image(x, parameters = c(0.5, 0.5, 0.5))
res
}
