% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/robust_brain_mask.R
\name{robust_brain_mask}
\alias{robust_brain_mask}
\title{Robust Brain Mask from Template}
\usage{
robust_brain_mask(file, template.file, template.mask = NULL, nvoxels = 7,
  ret_mask = FALSE, typeofTransform = "Rigid", verbose = TRUE, ...)
}
\arguments{
\item{file}{File for masking - either filename or class nifti}

\item{template.file}{Template to warp to original image space}

\item{template.mask}{Mask of template to use as rough brain mask.  If
\code{template.file} is specified, but \code{template.mask} is not,
then \code{fslbin(file=template.file)} is performed.}

\item{nvoxels}{Number of voxels in nxnxn direction to inflate, see
\code{\link{fslmaths.help}} for \code{-kernel boxv n -dilM}.}

\item{ret_mask}{Return mask of robust brain}

\item{typeofTransform}{Transformation for template to image, passed to
\code{\link{ants_regwrite}}.}

\item{verbose}{Print diagnostic progress}

\item{...}{not used}
}
\value{
Object of class nifti or vector of indices
}
\description{
Masks the brain from scan using a rigid body transformation
of the template mask to the native brain space and then inflating the mask
to adjust for mis-registration.
}

