% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_image.R
\name{smooth_image}
\alias{smooth_image}
\title{Gaussian smooth image using ANTsR}
\usage{
smooth_image(file, sigma = 10, mask = NULL, smooth_mask = TRUE,
  smoothed_mask = NULL, verbose = TRUE, retfile = FALSE, ...)
}
\arguments{
\item{file}{(character) image to be smoothed}

\item{sigma}{(numeric) sigma (in mm) of Gaussian kernel for smoothing}

\item{mask}{(character) optional mask given for image}

\item{smooth_mask}{(logical) Smooth mask?  If TRUE, the masked image 
will be divided by the smoothed mask.}

\item{smoothed_mask}{(character or antsImage) If specified and 
\code{smooth_mask = TRUE}, then will use this as the smoothed mask for 
division.}

\item{verbose}{(logical) print out command before running}

\item{retfile}{logical to indicate if an \code{antsImage} should be returned
(useful for chaining)}

\item{...}{additional arguments passed to \code{\link{smoothImage}}.}
}
\value{
Object of class \code{nifti}
}
\description{
This function calls \code{smoothIamge} to smooth an image and either
saves the image or returns an object of class nifti
}
\examples{
system.time({
library(oro.nifti)
dims = c(50, 50, 50)
x = array(rnorm(prod(dims)), dim = dims)
img = nifti(x, dim= dims,
datatype = convert.datatype()$FLOAT32, cal.min = min(x), 
cal.max = max(x), pixdim = rep(1, 4))
s.img = smooth_image(img, sigma = 3, sigmaInPhysicalCoordinates = TRUE)
})
}
