% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smri_preproc.R
\name{smri_preproc}
\alias{smri_preproc}
\title{Standard MRI Preprocessing}
\usage{
smri_preproc(files, window = FALSE, outfiles = NULL, maskfile = NULL,
  correct = TRUE, correction = "N4", correct_after_mask = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{files}{filenames (or nifti objects) of images to be processed.
Will register to the first scan}

\item{window}{Should \code{\link{robust_window}} before running?}

\item{outfiles}{(character) name of output files, with extension}

\item{maskfile}{Filename (or nifti object) of mask for image to be
registered to}

\item{correct}{do Bias field correction}

\item{correction}{(character) N3 or N4 correction?}

\item{correct_after_mask}{Should the inhomogeneity correction be run
after masking.}

\item{verbose}{Diagnostic messages}

\item{...}{arguments to \code{\link{preprocess_mri_within}}}
}
\value{
List from \code{\link{preprocess_mri_within}}
}
\description{
This function performs robust skull stripping on the 
T1, then preprocessing within visit
}
