% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ss_bias.R
\name{ss_bias}
\alias{ss_bias}
\title{Skull Strip then Bias Field Correct}
\usage{
ss_bias(filename, maskfile = NULL, outfile = filename, skull_strip = TRUE,
  bet.opts = "-B -f 0.1 -v", betcmd = "bet", n3correct = TRUE,
  correction = "N3", shrinkfactor = "4", verbose = TRUE, ...)
}
\arguments{
\item{filename}{filename to be processed}

\item{maskfile}{filename of masked to be passed in, otherwise bet is performed}

\item{outfile}{output filename with extension (should be .nii.gz}

\item{skull_strip}{should skull stripping be done}

\item{bet.opts}{options for fslbet}

\item{betcmd}{bet command to be used see \code{\link{fslbet}}}

\item{n3correct}{should n3 correction be done stripping be done}

\item{correction}{correction method used see \code{\link{bias_correct}}}

\item{shrinkfactor}{correction method used see \code{\link{n3BiasFieldCorrection}}}

\item{verbose}{Diagnostic output}

\item{...}{passed to \code{\link{bias_correct}}}
}
\description{
This function wraps \code{fslbet} and
\code{bias_correct}, performing skull stripping first
}
\seealso{
\code{\link{bias_ss}}
}

