% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_func.R
\name{stat_img}
\alias{stat_img}
\title{Create Statistics Image}
\usage{
stat_img(imgs, func = c("mean", "median", "mode", "sd", "var", "mad", "sum",
  "prod", "z"), finite = TRUE, ...)
}
\arguments{
\item{imgs}{Character vector, list of characters, or object of class 
\code{nifti}}

\item{func}{Function to perform voxel-wise on list of images}

\item{finite}{Should non-finite values be removed?}

\item{...}{Addictional arguments to pass to \code{func}}
}
\value{
Object of class \code{nifti}
}
\description{
Creates output from a list of images, performing voxel-wise
operations to create a statistic image
}

