% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sublime.R
\name{sublime}
\alias{sublime}
\title{OASIS Processing Pipeline}
\usage{
sublime(baseline_files, followup_files, baseline_outfiles, followup_outfiles,
  typeofTransform = "Rigid", skull_strip = FALSE,
  bet.opts = "-B -f 0.1 -v", betcmd = "bet", n3correct = FALSE,
  correction = "N3", shrinkfactor = "4",
  template.file = file.path(fsldir(), "data", "standard",
  "MNI152_T1_1mm_brain.nii.gz"), templateTransform = "Affine",
  verbose = TRUE, ...)
}
\arguments{
\item{typeofTransform}{type of transformed used, passed to
\code{\link{antsRegistration}}}

\item{skull_strip}{do skull stripping with FSL BET}

\item{bet.opts}{Options passed to \code{\link{fslbet}}}

\item{betcmd}{BET command used, passed to \code{\link{fslbet}}}

\item{n3correct}{do N3 Bias correction}

\item{template.file}{Filename of template to warp to}

\item{...}{arguments to \code{\link{whitestripe}}}

\item{filename}{filename of T1 image}

\item{skull_stripfile}{Output skull strip filename}

\item{normalize}{Normalize data using \code{\link{whitestripe}}}

\item{normalize_file}{\code{\link{whitestripe}} image mask}

\item{retimg}{return a nifti object from function}

\item{outfile}{output filename should have .nii or .nii.gz
extension}

\item{interpolator}{interpolation done for
\code{\link{antsApplyTransforms}}}

\item{other.files}{Filenames of other iamges to be
transformed with the T1}

\item{other.outfiles}{Output filenames of \code{other.files} to
be written}

\item{remove.warp}{(logical) Should warping images be deleted?}

\item{outprefix}{Character path of where the warp files should be stored.
Required if \code{remove.warp = FALSE}}
}
\value{
NULL or object of class nifti for transformed T1 image
}
\description{
This function performs registration to a T1 template
using ANTsR and SyN transformation
}

