% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_4d.R
\docType{methods}
\name{subset_4d-methods}
\alias{.subset_4d}
\alias{subset_4d}
\alias{subset_4d,antsImage-method}
\alias{subset_4d,character-method}
\alias{subset_4d,nifti-method}
\alias{subset_4d-methods}
\title{Subset a 4D image}
\usage{
subset_4d(img, ind, ...)

\S4method{subset_4d}{character}(img, ind, ...)

\S4method{subset_4d}{nifti}(img, ind, ...)

\S4method{subset_4d}{antsImage}(img, ind, ...)

.subset_4d(img, ind)
}
\arguments{
\item{img}{character path of image or 
an object of class nifti, or antsImage}

\item{ind}{subset indexers for the 4th dimension (usually time)}

\item{...}{Arguments passed to \code{\link{copyNIfTIHeader}} or 
\code{\link{antsCopyImageInfo}}}
}
\value{
An \code{\link{antsImage}} or \code{\link{nifti}} depending on
input
}
\description{
Subsets an \code{\link{antsImage}} or \code{\link{nifti}}
object from 4D
}
\examples{
library(oro.nifti)
n = 20
x = nifti(array(rnorm(n^3*10), dim = c(n, n, n, 10)))
ind = 2:3
subset_4d(x, ind)
subset_4d(x, ind = 1) 
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}

