% Generated by roxygen2 (4.0.2): do not edit by hand
\name{t1_syn}
\alias{t1_syn}
\title{Registration to T1 Template}
\usage{
t1_syn(filename, n3correct = FALSE, retimg = TRUE, outfile = NULL,
  interpolator = "Linear", other.files = NULL, other.outfiles = NULL,
  native.cereb = TRUE, native.fname = NULL, atlas.file = NULL,
  typeofTransform = "SyN", ...)
}
\arguments{
\item{filename}{filename of T1 image}

\item{n3correct}{do N3 Bias correction}

\item{retimg}{return a nifti object from function}

\item{outfile}{output filename should have .nii or .nii.gz
extension}

\item{interpolator}{interpolation done for
\code{\link{antsApplyTransforms}}}

\item{other.files}{Filenames of other iamges to be
transformed with the T1}

\item{other.outfiles}{Output filenames of \code{other.files} to
be written}

\item{native.cereb}{Logical indicating if native cerebellum should be
created to \code{native.fname}}

\item{native.fname}{filename of native cerebellum file}

\item{atlas.file}{Filename of atlas used for warping}

\item{typeofTransform}{type of transformed used, passed to
\code{\link{antsRegistration}}}

\item{...}{arguments to \code{\link{antsApplyTransforms}}}
}
\value{
NULL or object of class nifti for transformed T1 image
}
\description{
This function performs registration to a T1 template
using ANTsR and SyN transformation
}

