% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformlist_from_outprefix.R
\name{transformlist_from_outprefix}
\alias{transformlist_from_outprefix}
\title{Get list of transforms that would likely be made with a specific `outprefix`}
\usage{
transformlist_from_outprefix(outprefix)
}
\arguments{
\item{outprefix}{If you used this in \code{\link{antsRegistration}}, then
what would be the filenames of the transforms}
}
\value{
A list of character vectors
}
\description{
Get list of transforms that would likely be made with a specific `outprefix`
}
\examples{
outprefix = tempfile()
transformlist_from_outprefix(outprefix)
}
