% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/within_visit_registration.R
\name{within_visit_registration}
\alias{within_visit_registration}
\title{OASIS Processing Pipeline}
\usage{
within_visit_registration(fixed, moving, outfiles, typeofTransform = "Rigid",
  ...)
}
\arguments{
\item{fixed}{filename of fixed image to be registered to.}

\item{moving}{filenames (or nifti) of images to register to fixed image}

\item{typeofTransform}{Transformation of moving to fixed image}

\item{...}{additional arguments to \code{\link{ants_regwrite}}}
}
\description{
This function performs registration to a T1 template
using ANTsR and SyN transformation
}

