% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transformlist.R
\name{write_transformlist}
\alias{write_transformlist}
\title{Read Transform List from ANTsR}
\usage{
write_transformlist(transformlist)
}
\arguments{
\item{transformlist}{List of transform elements}
}
\value{
List of transforms
}
\description{
Transform a transformlist from ANTsR into an object in R
}

