% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_pad.R
\name{zero_pad}
\alias{zero_pad}
\title{Zero pads an image}
\usage{
zero_pad(img, ...)
}
\arguments{
\item{img}{Array, class \code{\link{nifti}}, or 
\code{antsImage}}

\item{...}{Options to \code{\link[neurobase]{zero_pad}}}
}
\value{
Object of class nifti, array, or \code{antsImage}
}
\description{
Zero pads an image
}
