% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_template.R
\name{zscore_template}
\alias{zscore_template}
\title{Create Z-score-to-Template Image}
\usage{
zscore_template(img, template.file, mean.img, sd.img, outfile = NULL,
  typeofTransform = "SyN", interpolator = "Linear", robust = TRUE,
  native = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{img}{Image to be z-scored}

\item{template.file}{Template file to be registered to.  An affine registration
will be done usually, so if the image is skull-stripped, use a skull-stripped
template and vice versa}

\item{mean.img}{Mean image to be subtracted}

\item{sd.img}{SD image to be divded}

\item{outfile}{Output filename.  Will use \code{\link{writenii}} to write}

\item{typeofTransform}{Transform for registration.  See \code{\link{antsRegistration}}}

\item{interpolator}{Interpolator to use, See \code{\link{antsRegistration}}}

\item{robust}{Should See \code{\link{robust_window}} be used on the image}

\item{native}{Should a native-space image (default) or template-space image
be returned}

\item{verbose}{Print diagnostic output}

\item{...}{Options passed to \code{\link{robust_window}}.}
}
\value{
Object of class \code{nifti}
}
\description{
Registers an image to a template, subtracts a mean and sd image, and
will return the image either in template space or native.
}

