% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oro2ants.R
\name{ants2oro}
\alias{ants2oro}
\title{Convert Between antsImage and nifti objects}
\usage{
ants2oro(img, reorient = FALSE, reference = NULL, ..., cleanup = TRUE,
  drop_dim = TRUE)
}
\arguments{
\item{img}{Object of class antsImage}

\item{reorient}{Reorientation passed to \code{\link{readnii}}}

\item{reference}{Object of class \code{\link{nifti}} to copy header
information}

\item{...}{Arguments passed to \code{\link{copyNIfTIHeader}} if 
reference is set and \code{img} is \code{antsImage}}

\item{cleanup}{temporary files are deleted after they are read in}

\item{drop_dim}{Should \code{\link{drop_img_dim}} be run after reading? 
Passed to \code{\link{readnii}}}
}
\value{
Object of class \code{nifti}.
}
\description{
NIfTI data can be converted between \code{antsImage} 
(from the ANTsR package) and nifti S4 objects.
}
