% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_apply_transforms.R
\docType{methods}
\name{ants_apply_transforms-methods}
\alias{ants_apply_transforms}
\alias{ants_apply_transforms,ANY,ANY,"ANY"-method}
\alias{ants_apply_transforms,ANY,ANY-method}
\alias{ants_apply_transforms,ANY,character,ANY-method}
\alias{ants_apply_transforms,ANY,character-method}
\alias{ants_apply_transforms,ANY,list,ANY-method}
\alias{ants_apply_transforms,ANY,list-method}
\alias{ants_apply_transforms-methods}
\title{Apply ANTs transforms to nifti images}
\usage{
ants_apply_transforms(fixed, moving, transformlist, ...)

\S4method{ants_apply_transforms}{ANY,list}(fixed, moving, transformlist,
  ret_ants = FALSE, ...)

\S4method{ants_apply_transforms}{ANY,character}(fixed, moving, transformlist,
  ret_ants = FALSE, ...)

\S4method{ants_apply_transforms}{ANY,ANY}(fixed, moving, transformlist,
  ret_ants = FALSE, ...)
}
\arguments{
\item{fixed}{image defining domain into which the moving image is transformed.}

\item{moving}{image to be mapped to fixed space.}

\item{transformlist}{character vector of transforms generated by 
\code{\link{antsRegistration}}}

\item{...}{Arguments to be passed to \code{\link{antsApplyTransforms}}.}

\item{ret_ants}{if \code{TRUE}, the returned object is an \code{antsImage},
as opposed to a \code{nifti}}
}
\value{
Either \code{nifti} object or list of nifti objects
}
\description{
Wraps \code{\link{antsApplyTransforms}} but allows for 
\code{\link{nifti}} objects from the \code{oro.nifti} package
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}
\seealso{
\code{\link{antsApplyTransforms}}
}

