% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ants.R
\docType{methods}
\name{check_ants,nifti-method}
\alias{check_ants,nifti-method}
\title{Check if antsImage or read in}
\usage{
\S4method{check_ants}{nifti}(x)
}
\value{
antsImage object
}
\description{
Simple check to see if input is character, list, \code{nifti},
or class \code{antsImage}
}
