% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkimg_antsImage.R
\docType{methods}
\name{checkimg-methods}
\alias{checkimg-methods}
\alias{checkimg,antsImage-method}
\title{Force object to filename}
\usage{
\S4method{checkimg}{antsImage}(file, ...)
}
\arguments{
\item{file}{character or \code{nifti} object}

\item{...}{options passed to \code{\link{tempimg}}}
}
\description{
Ensures the output to be a character filename (or vector) from an input
image or \code{nifti}.
}
