% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_self.R
\name{diff_self}
\alias{diff_self}
\title{Flipped Difference Image}
\usage{
diff_self(img, swapdim = TRUE, verbose = TRUE)
}
\arguments{
\item{img}{Object of class \code{nifti}, \code{ANTsR}, or \code{character}}

\item{swapdim}{Should the image be reoriented to RPI then flipped, registered,
then changed back?}

\item{verbose}{Print diagnostic messages}
}
\value{
Object of class \code{nifti}
}
\description{
This function flips the image over the left/right axis, then registers 
the flipped image to the original image and then takes a difference
}

