% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filler.R
\name{filler}
\alias{filler}
\title{Image Mask Filler}
\usage{
filler(img, fill_size = 7, verbose = TRUE)
}
\arguments{
\item{img}{Object of class \code{\link{nifti}} or \code{antsImage}}

\item{fill_size}{Size of fill (in voxels)}

\item{verbose}{Output diagnostic values}
}
\value{
Filled object with same class as \code{img}
}
\description{
Fills in a binary image using \code{\link{iMath}}, 
using dilation erosion but with zero padding built in.
}

