% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_similarity.R
\name{image_similarity}
\alias{image_similarity}
\title{Compute Image Similarity Metrics}
\usage{
image_similarity(image1, image2, image1_mask = NULL, image2_mask = NULL, ...)
}
\arguments{
\item{image1}{Object of class \code{nifti} or character filename.
Passed to the \code{fixed} image in \code{\link{imageSimilarity}}}

\item{image2}{Object of class \code{nifti} or character filename
Passed to the \code{moving} image in \code{\link{imageSimilarity}}}

\item{image1_mask}{Mask for first image (optional)}

\item{image2_mask}{Mask for second image (optional)}

\item{...}{Additional arguments to pass to
\code{\link{imageSimilarity}}}
}
\value{
A numeric value
}
\description{
Compute Image Similarity Metrics
}
\note{
Not all metrics are symmetric.
}
\examples{
library(ANTsR)
library(extrantsr)
x =  getANTsRData( 'r16' )
y =  getANTsRData( 'r30' )
image_similarity(x,y, type="MeanSquares")
image_similarity(y,x, type="MeanSquares")

image_similarity(x,y, type="Correlation")
image_similarity(y,x, type="Correlation")

# NOT SYMMETRIC
image_similarity(x,y, type="MattesMutualInformation")
image_similarity(y,x, type="MattesMutualInformation")

# NOT SYMMETRIC
mutual_information(x, y)
mutual_information(y, x)

image_similarity(x,y, type="ANTSNeighborhoodCorrelation")
image_similarity(y,x, type="ANTSNeighborhoodCorrelation")
image_similarity(x,y, type="JointHistogramMutualInformation")
image_similarity(y,x, type="JointHistogramMutualInformation")
image_similarity(x,y, type="Demons")
image_similarity(y,x, type="Demons")
}
