% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/non_exported.R
\name{makepng}
\alias{makepng}
\title{Non-exported function to make png}
\usage{
makepng(outfile, addstub, ...)
}
\arguments{
\item{outfile}{Filename to be added to for png}

\item{addstub}{extension to add to outfile}

\item{...}{arguments passed to \code{\link{png}}}
}
\value{
Alpha level depending on device
}
\description{
Make my png
}

