% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malf.R
\name{malf}
\alias{malf}
\title{Multi-Atlas Label Fusion}
\usage{
malf(infile, template.images, template.structs, inverted = FALSE,
  keep_images = FALSE, outfiles = NULL, outfile = NULL, retimg = TRUE,
  func = "mode", ties.method = "first", keep_regs = FALSE,
  outprefix = NULL, interpolator = "NearestNeighbor",
  typeofTransform = "SyN", verbose = TRUE, ...)
}
\arguments{
\item{infile}{Input Image file}

\item{template.images}{Template Images to register 
to \code{infile}}

\item{template.structs}{Template gold standards to apply 
registration into \code{infile} space}

\item{inverted}{Should the MALF be inverted 
(infile to template then use inverse transforms)}

\item{keep_images}{Keep the \code{template.structs} in 
\code{infile} space}

\item{outfiles}{Output filenames for  \code{template.structs} in 
\code{infile} space}

\item{outfile}{Fused output filename}

\item{retimg}{Return Image to user using \code{\link{readnii}}}

\item{func}{function to combine labels.  See \code{\link{stat_img}}.}

\item{ties.method}{If \code{func = "mode"}, then this is passed to 
\code{\link{stat_img}}.}

\item{keep_regs}{Keep list of registrations.  If \code{TRUE}, then
\code{remove.warp = FALSE} in \code{\link{registration}}}

\item{outprefix}{passed to \code{\link{registration}} if 
\code{keep_regs = TRUE}}

\item{interpolator}{interpolation done for 
\code{\link{antsApplyTransforms}}}

\item{typeofTransform}{type of transformed used, passed to 
\code{\link{antsRegistration}}}

\item{verbose}{Print diagnostic output}

\item{...}{Arguments to be passed to \code{\link{malf_registration}}, which
really are options for \code{\link{registration}}}
}
\value{
The output filename or the nifti image or list of registrations and
output file
}
\description{
Takes in an input file and template images with 
a set of template structures and creates a label fusion
}
