% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_reg.R
\name{multi_reg}
\alias{multi_reg}
\title{Multi-Atlas Label Fusion}
\usage{
multi_reg(infiles, template.file, outfiles = NULL, typeofTransform = "SyN",
  interpolator = "Linear", outprefix = NULL, retimg = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{outfiles}{Output filenames for  \code{template.structs} in 
\code{infile} space}

\item{retimg}{Return Image to user using \code{\link{readNIfTI}}}

\item{verbose}{Print diagnostic output}

\item{...}{Arguments to be passed to \code{\link{ants_regwrite}}}

\item{infile}{Input Image file}

\item{template.images}{Template Images to register 
to \code{infile}}

\item{template.structs}{Template gold standards to apply 
registration into \code{infile} space}

\item{keep_images}{Keep the \code{template.structs} in 
\code{infile} space}

\item{outfile}{Fused output filename}
}
\value{
The output filename or the nifti image
}
\description{
Takes in an input file and template images with 
a set of template structures and creates a label fusion
}

