% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/non_exported.R
\name{my.tab}
\alias{my.tab}
\title{Quick Tabulation for logical vectors}
\usage{
my.tab(x, y, dnames = c("x", "y"))
}
\arguments{
\item{x}{Logical or 0/1 vector}

\item{y}{Logical or 0/1 vector}

\item{dnames}{names for table}
}
\value{
table of x vs y
}
\description{
Speeds up on table for logical x and y
}

