% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oro2ants.R
\name{oro2ants}
\alias{oro2ants}
\title{Convert Between nifti and antsImage}
\usage{
oro2ants(img, reference = NULL, cleanup = TRUE, ...)
}
\arguments{
\item{img}{Object of class \code{nifti}}

\item{reference}{Object of class \code{antsImage} to 
copy header information (origin, spacing, direction) (experimental)}

\item{cleanup}{temporary files are deleted after they are read in}

\item{...}{arguments passed to \code{\link{checkimg}}}
}
\value{
Object of class \code{antsImage}
}
\description{
NIfTI data can be converted between \code{nifti} 
(from the oro.nifti package) and \code{antsImage} objects.
}
\examples{
arr = array(rnorm(20^3), dim = rep(20, 3))
img = oro.nifti::nifti(arr)
oro2ants(img)
}
