% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/oro2ants.R
\name{oro2ants}
\alias{oro2ants}
\title{Convert Between oro.nifti and antsImage}
\usage{
oro2ants(img, reorient = FALSE)
}
\arguments{
\item{img}{Object of class \code{nifti}}

\item{reorient}{Reorientation passed to \code{\link{readNIfTI}}}
}
\value{
Object of class \code{antsImage}
}
\description{
NIfTI data can be converted between \code{nifti}
(from the oro.nifti package) and \code{antsImage} objects.
}

