% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sublime.R
\name{preprocess_mri_across}
\alias{preprocess_mri_across}
\title{OASIS Processing Pipeline}
\usage{
preprocess_mri_across(baseline_files, followup_files, baseline_outfiles,
  followup_outfiles, retimg = FALSE, reorient = FALSE, maskfile = NULL,
  skull_strip = FALSE, bet.opts = "-B -f 0.1 -v", betcmd = "bet",
  within.transform = "Rigid", within.interpolator = "LanczosWindowedSinc",
  across.transform = within.transform,
  across.interpolator = within.interpolator, verbose = TRUE, ...)
}
\arguments{
\item{retimg}{return a nifti object from function}

\item{skull_strip}{do skull stripping with FSL BET}

\item{bet.opts}{Options passed to \code{\link{fslbet}}}

\item{betcmd}{BET command used, passed to \code{\link{fslbet}}}

\item{...}{arguments to \code{\link{whitestripe}}}

\item{filename}{filename of T1 image}

\item{skull_stripfile}{Output skull strip filename}

\item{n3correct}{do N3 Bias correction}

\item{normalize}{Normalize data using \code{\link{whitestripe}}}

\item{normalize_file}{\code{\link{whitestripe}} image mask}

\item{outfile}{output filename should have .nii or .nii.gz
extension}

\item{template.file}{Filename of template to warp to}

\item{interpolator}{interpolation done for
\code{\link{antsApplyTransforms}}}

\item{other.files}{Filenames of other iamges to be
transformed with the T1}

\item{other.outfiles}{Output filenames of \code{other.files} to
be written}

\item{typeofTransform}{type of transformed used, passed to
\code{\link{antsRegistration}}}

\item{remove.warp}{(logical) Should warping images be deleted?}

\item{outprefix}{Character path of where the warp files should be stored.
Required if \code{remove.warp = FALSE}}
}
\value{
NULL or list of objects of class nifti for transformed images
}
\description{
This function performs registration to a T1 template
using ANTsR and SyN transformation
}

