% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapply_malf.R
\name{reapply_malf}
\alias{reapply_malf}
\title{Reapply MALF Operation}
\usage{
reapply_malf(infile, regs, template.structs, keep_images = FALSE,
  outfiles = NULL, outfile = NULL, retimg = TRUE, func = "mode",
  ties.method = "first", interpolator = NULL, verbose = TRUE, ...)
}
\arguments{
\item{infile}{Input Image file}

\item{template.structs}{Template gold standards to apply 
registration into \code{infile} space}

\item{outfiles}{Output filenames for  \code{template.structs} in 
\code{infile} space}

\item{interpolator}{interpolation done for 
\code{\link{antsApplyTransforms}}}

\item{verbose}{Print diagnostic output}

\item{...}{Arguments to be passed to \code{\link{ants_apply_transforms}}}
}
\value{
The output filename or the nifti image or list of registrations and
output file
}
\description{
Takes in an input file, set of registrations, 
and a set of template structures and applies the registrations,
with a different interpolator if necessary, and then combines output
}

