% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_whitestripe.R
\name{reg_whitestripe}
\alias{reg_whitestripe}
\title{WhiteStripe Normalization}
\usage{
reg_whitestripe(t1 = NULL, t2 = NULL, register = TRUE, native = TRUE,
  template.file = file.path(fsldir(), "data", "standard",
  paste0("MNI152_T1_1mm", ifelse(is.null(mask), "", "_brain"), ".nii.gz")),
  typeofTransform = c("Rigid", "Affine"),
  interpolator = "LanczosWindowedSinc", type = c("T1", "T2", "hybrid"),
  t1.outfile = NULL, t2.outfile = NULL, other.files = NULL,
  other.outfiles = NULL, ws.outfile = NULL, mask = NULL,
  mask.outfile = NULL, verbose = TRUE, ...)
}
\arguments{
\item{t1}{filename (or nifti objects) of T1 image}

\item{t2}{filename (or nifti objects) of T2 image}

\item{register}{Register image to template file}

\item{native}{If images are registered, should the native space
normalized image be returned?}

\item{template.file}{Filename of template to warp to}

\item{typeofTransform}{type of transformed used, passed to 
\code{\link{antsRegistration}}}

\item{interpolator}{Interpolation to be performed, passed to
\code{\link{antsRegistration}}}

\item{type}{Type of whitestripe normalization done}

\item{t1.outfile}{Output filename of normalized T1 image}

\item{t2.outfile}{Output filename of normalized T2 image}

\item{other.files}{Character filenames or list of nifti objects to 
normalize.  In the same space as T1.}

\item{other.outfiles}{Character filenames for output 
normalized files.}

\item{ws.outfile}{Character filename for output 
whitestripe mask.}

\item{mask}{File or nifti image of brain mask}

\item{mask.outfile}{Character filename for output 
brain mask.}

\item{verbose}{Print Diagnostic Messages}

\item{...}{arguments to \code{\link{whitestripe}} or 
\code{\link{whitestripe_hybrid}}}
}
\value{
List of nifti objects or character filenames
}
\description{
This function performs registration to a T1 template
using ANTsR and SyN transformation
}

