% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_neck}
\alias{remove_neck}
\title{Remove Neck from CT Scan}
\usage{
remove_neck(file, template.file = system.file("scct_unsmooth.nii.gz", package
  = "cttools"),
  template.mask = system.file("scct_unsmooth_Skull_Stripped_Mask.nii.gz",
  package = "cttools"), ret_mask = FALSE, rep.value = 0)
}
\arguments{
\item{file}{File for neck removal - either filename or class nifti}

\item{template.file}{Template to warp to original image space}

\item{template.mask}{Mask of template to use as rough brain mask}

\item{ret_mask}{Return mask of slices to keep}

\item{rep.value}{Value to replace neck slices with}
}
\value{
Object of class nifti or vector of indices
}
\description{
Removes the neck from axially acquired CT scan so
Skull stripping can be done.
}

