% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/remove_neck.R
\name{remove_neck}
\alias{remove_neck}
\title{Remove Neck from CT Scan}
\usage{
remove_neck(file, template.file = system.file("scct_unsmooth.nii.gz", package
  = "cttools"),
  template.mask = system.file("scct_unsmooth_Skull_Stripped_Mask.nii.gz",
  package = "cttools"), ret_mask = FALSE, typeofTransform = "Rigid",
  rep.value = 0, ...)
}
\arguments{
\item{file}{File for neck removal - either filename or class nifti}

\item{template.file}{Template to warp to original image space}

\item{template.mask}{Mask of template to use as rough brain mask}

\item{ret_mask}{Return mask of slices to keep}

\item{typeofTransform}{Transformation for template to image, passed to
\code{\link{ants_regwrite}}.}

\item{rep.value}{Value to replace neck slices with}

\item{...}{not used}
}
\value{
Object of class nifti or vector of indices
}
\description{
Removes the neck from axially acquired CT scan so
Skull stripping can be done.
}

