% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_neck.R
\name{remove_neck}
\alias{remove_neck}
\title{Remove Neck from CT Scan}
\usage{
remove_neck(file, template.file, template.mask = NULL, ret_mask = FALSE,
  typeofTransform = "Rigid", rep.value = 0, verbose = TRUE, ...)
}
\arguments{
\item{file}{File for neck removal - either filename or class nifti}

\item{template.file}{Template to warp to original image space}

\item{template.mask}{Mask of template to use as rough brain mask.  If
\code{template.file} is specified, but \code{template.mask} is not,
then \code{fslbin(file=template.file)} is performed.}

\item{ret_mask}{Return mask of slices to keep}

\item{typeofTransform}{Transformation for template to image, passed to
\code{\link{ants_regwrite}}.}

\item{rep.value}{Value to replace neck slices with}

\item{verbose}{Print out diagnostic messages}

\item{...}{not used}
}
\value{
Object of class nifti or vector of indices
}
\description{
Removes the neck from axially acquired CT scan so
Skull stripping can be done.
}

