% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_neck.R
\name{remove_neck}
\alias{remove_neck}
\title{Remove Neck from Image}
\usage{
remove_neck(file, template.file, template.mask = NULL,
  ret_mask = FALSE, typeofTransform = "Rigid", rep.value = 0,
  swapdim = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{file}{File for neck removal - either filename or class nifti}

\item{template.file}{Template to warp to original image space}

\item{template.mask}{Mask of template to use as rough brain mask.  If
\code{template.file} is specified, but \code{template.mask} is not,
then \code{fslbin(file=template.file)} is performed.}

\item{ret_mask}{Return mask of slices to keep}

\item{typeofTransform}{Transformation for template to image, passed to
\code{\link{ants_regwrite}}.}

\item{rep.value}{Value to replace neck slices with}

\item{swapdim}{Should the dimensions be swapped before registration,
and then reset after}

\item{verbose}{Print out diagnostic messages}

\item{...}{Additional arguments passed to \code{\link{ants_regwrite}}}
}
\value{
Object of class nifti or vector of indices
}
\description{
Removes the neck from axially acquired scan so
skull stripping can be done.
}
\examples{
if (fslr::have.fsl()){
  print(fslr::fsl_version())
  in_ci <- function() {
    nzchar(Sys.getenv("CI"))
  }
  if (in_ci()) {
    destfile = tempfile(fileext = ".nii.gz")
    dl = download.file(paste0(
      "https://github.com/muschellij2/",
      "Neurohacking/files/3454385/113-01-MPRAGE2.nii.gz"),
      destfile = destfile)
    tfile = fslr::mni_fname(brain = TRUE)
    if (file.exists(tfile)) {
      res = remove_neck(
        destfile, 
        template.file = tfile,
        template.mask = fslr::mni_fname(brain = TRUE, mask = TRUE)
      )
    }
  }
}
}
