% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/non_exported.R
\name{sim}
\alias{sim}
\title{Calculate Overlap Metrics}
\usage{
sim(dman, dauto, dim.dman, dim.dauto)
}
\arguments{
\item{dman}{vector of manually delineated values}

\item{dauto}{vector of automatically delineated values}

\item{dim.dman}{dimension of mask for manual}

\item{dim.dauto}{dimension of mask for automatic}
}
\value{
List of performance measures
}
\description{
Calculates perofrmance metrics, like dice, Jaccard,
sensitivity, specificity
}

