% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_exported.R
\name{sim}
\alias{sim}
\title{Calculate Overlap Metrics}
\usage{
sim(dman, dauto)
}
\arguments{
\item{dman}{vector of manually delineated values}

\item{dauto}{vector of automatically delineated values}
}
\value{
List of performance measures
}
\description{
Calculates perofrmance metrics, like dice, Jaccard,
sensitivity, specificity
}
