% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_image.R
\name{smooth_image}
\alias{smooth_image}
\title{Gaussian smooth image using ANTsR}
\usage{
smooth_image(file, sigma = 10, mask = NULL, smooth_mask = TRUE,
  verbose = TRUE, ...)
}
\arguments{
\item{file}{(character) image to be smoothed}

\item{sigma}{(numeric) sigma (in mm) of Gaussian kernel for smoothing}

\item{mask}{(character) optional mask given for image}

\item{smooth_mask}{(logical) Smooth mask?  If TRUE, the masked image 
will be divided by the smoothed mask.}

\item{verbose}{(logical) print out command before running}

\item{...}{additional arguments passed to \code{\link{smoothImage}}.}
}
\value{
Object of class \code{nifti}
}
\description{
This function calls \code{smoothIamge} to smooth an image and either
saves the image or returns an object of class nifti
}
\examples{
system.time({
library(oro.nifti)
x = array(rnorm(1e6), dim = c(100, 100, 100))
img = nifti(x, dim= c(100, 100, 100), 
datatype = convert.datatype()$FLOAT32, cal.min = min(x), 
cal.max = max(x), pixdim = rep(1, 4))
s.img = smooth_image(img, sigmaInPhysicalCoordinates = TRUE)
})
}

