% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_func.R
\name{stat_img}
\alias{stat_img}
\title{Create Statistics Image}
\usage{
stat_img(imgs, func = c("mean", "median", "mode", "peak", "sd", "var", "mad",
  "sum", "prod", "z", "quantile"), finite = TRUE, masks = NULL,
  na_masks = TRUE, ...)
}
\arguments{
\item{imgs}{Character vector, list of characters, or object of class
\code{nifti}}

\item{func}{Function to perform voxel-wise on list of images}

\item{finite}{Should non-finite values be removed?}

\item{masks}{Character vector, list of characters, or object of class
\code{nifti} same length as images.}

\item{na_masks}{if \code{masks} are given, should values of \code{0} be
turned into \code{NA}?}

\item{...}{Addictional arguments to pass to \code{func}}
}
\value{
Object of class \code{nifti}
}
\description{
Creates output from a list of images, performing voxel-wise
operations to create a statistic image
}
\note{
When \code{func = "mode"}, the data is tabulated and then
\code{\link{max.col}} is run.  The user can pass \code{ties.method} to
determine how to break ties.  The default is \code{"first"} as compared to
\code{\link{max.col}} where it is \code{"random"}.  When \code{func = "peak"},
\code{\link{density}} is run on each voxel and the value with the maximum.  If
the number of unique values is only 1, that value is returned.
}
