% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempants.R
\name{tempants}
\alias{tempants}
\title{Create temporary nii.gz file from antsImage}
\usage{
tempants(x, gzipped = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{antsImage}}

\item{gzipped}{logical if the tempfile should be gzipped or not}

\item{...}{additional arguments to pass to \code{\link{tempimg}}
if \code{nifti} object}
}
\value{
Character value of filename
}
\description{
Takes in a object of class \code{antsImage},
writes it to a temp file, and appends .nii or .nii.gz
}
